### 2018-07-30 #864 @whedon
[PRE REVIEW]: Yellowbrick: Visualizing the Scikit-Learn Model Selection Process
**Submitting author:** @bbengfort (<a href="http://orcid.org/0000-0003-0660-7682">Benjamin Bengfort</a>)
**Repository:** <a href="https://github.com/DistrictDataLabs/yellowbrick" target ="_blank">https://github.com/DistrictDataLabs/yellowbrick</a>
**Version:** v0.8
