### 2018-07-30 #864 @arfon
@jakevdp - the submitting author suggested you as the handling editor.
