### 2018-07-30 #864 @arfon
@bbengfort - looks like there's some issues with your `paper.bib` file. Can you take a look at fixing it up?

You can test your changes with:
