### 2018-07-30 #864 @bbengfort
The proof looks good, there are a couple changes that I think might tighten this up:

1. Change title to "Yellowbrick: Visualizing the Model Selection Process" to get the title on one line.
2. Instead of 5 separate figures, I can generate one big figure, unless there is a way to control placement in a Markdown file like in a TeX file.

Looking forward to your review!
