### 2018-07-30 #864 @rebeccabilbro
Yes, @jakevdp - looking forward to your feedback!
