### 2018-07-31 73d9e88
Add docstring and quick method
