+++ yellowbrick/target/feature_correlation.py
@@ -1,11 +1,10 @@
 # yellowbrick.classifier.feature_correlation
-# Class balance visualizer for showing per-class support.
+# Feature correlation to dependent variable visualizer.
 #
 # Author    Zijie (ZJ) Poh <poh.zijie@gmail.com>
-# Author:   Benjamin Bengfort <bbengfort@districtdatalabs.com>
 # Created:  Wed Jul 29 15:30:40 2018 -0700
 #
-# ID: feature_correlation.py [5388065] neal@nhumphrey.com $
+# ID: feature_correlation.py [] poh.zijie@gmail.com $
 
 """
 Feature Correlation to Dependent Variable Visualizer.
@@ -17,7 +16,9 @@ Feature Correlation to Dependent Variable Visualizer.
 
 import numpy as np
 
-from .base import TargetVisualizer
+from yellowbrick.target.base import TargetVisualizer
+from yellowbrick.utils import is_dataframe
+from yellowbrick.exceptions import YellowbrickValueError
 
 from sklearn.feature_selection import (mutual_info_classif,
                                        mutual_info_regression)
@@ -29,48 +30,167 @@ from scipy.stats import pearsonr
 ##########################################################################
 
 class FeatureCorrelation(TargetVisualizer):
+    """
+    Displays the correlation between features and dependent variables.
 
-    method_func = {
-        'pearson': lambda X, y: [pearsonr(x, y)[0] for x in X.T],
+    This visualizer can be used side-by-side with
+    yellowbrick.features.JointPlotVisualizer that plots a feature
+    against the target and shows the distribution of each via a
+    histogram on each axis.
+
+    Parameters
+    ----------
+    X : ndarray or DataFrame of shape n x m
+        A matrix of n instances with m features
+
+    y : ndarray or Series of length n
+        An array or series of target or class values
+
+    ax : matplotlib Axes, default: None
+        The axis to plot the figure on. If None is passed in the current axes
+        will be used (or generated if required).
+
+    method : one of {'pearson', 'mutual_info'}, default: 'pearson'
+        The method to calculate correlation between features and target.
+
+    classification : boolean, default: False
+        This parameter is only used for method = 'mutual_info'.
+
+    labels : list, default: None
+        A list of feature names to use. If a DataFrame is passed to fit and
+        features is None, feature names are selected as the column names.
+
+    sort : boolean, default: False
+        If false, the features are are not sorted in the plot; otherwise
+        features are sorted in ascending order of correlation.
+
+    feature_index : ndarray or a list of feature names
+        An array of feature index or feature names to include in the plot.
+        If a list of feature names is provided, then X must be a DataFrame.
+
+    kwargs : dict
+        Keyword arguments that are passed to the base class and may influence
+        the visualization as defined in other Visualizers.
+
+    Attributes
+    ----------
+    features_ : np.array
+        The feature labels
+
+    corr_ : np.array
+        Correlation between features and dependent variable.
+
+    Examples
+    --------
+
+    >>> viz = FeatureCorrelation()
+    >>> viz.fit(X, y)
+    >>> viz.poof()
+    """
+
+    METHOD_FUNC = {
+        'pearson': lambda X, y: [pearsonr(x, y)[0] for x in np.asarray(X).T],
         'mutual_info': lambda X, y: mutual_info_regression(X, y),
         'mutual_info_classif': lambda X, y: mutual_info_classif(X, y)
     }
-    method_label = {
+    METHOD_LABEL = {
         'pearson': 'Pearson Correlation',
         'mutual_info': 'Mutual Information',
         'mutual_info_classif': 'Mutual Information'
     }
 
     def __init__(self, ax=None, method='pearson', classification=False,
[…]
