### 2018-07-31 c3b238c
[#529] Add links to scoring metrics documentation
In the `KElbowVisualizer` docstring, I added links to scikit-learn's
documentation for the `silhouette_score` and `calinski_harabaz_score`
functions. This is to help distinguish them from the visualizer's parameter
values `silhouette` and `calinksi_harabaz`.

In addition, I included a link to the documentation for the default scoring
function, `distortion_score`, which is implemented in Yellowbrick.

Fixes #529
