### 2018-07-31 51fb849
update tests per PR comments
