### 2018-07-31 023d61e
adds indents to the seealso in the docstring so that sphinx won't complain
