+++ yellowbrick/datasets/base.py
@@ -17,214 +17,224 @@ Loading utilities for the yellowbrick datasets.
 ##########################################################################
 
 import os
+import json
 import numpy as np
 
-from sklearn.datasets.base import Bunch
+from .download import download_data
+from .path import find_dataset_path, dataset_exists
 
-from .path import find_dataset_path, FIXTURES
+from yellowbrick.utils.decorators import memoized
 
-
-##########################################################################
-## Links and MD5 hash of datasets
-##########################################################################
-
-DATASETS = {
-    'concrete': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/concrete.zip',
-        'signature': 'b9ea5f26a7bb272a040e2f1a993b26babbf8dc4a04ab8198bb315ca66d71f10d',
-        'type': 'numpy',
-    },
-    'energy': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/energy.zip',
-        'signature': '19fb86f3bcdde208eed46944172cb643ef6a7d58da103fb568fae43205ed89d3',
-        'type': 'numpy',
-    },
-    'credit': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/credit.zip',
-        'signature': '4a91339c69f55e18f3f48004328fbcb7868070b618208fed099920427b084e5e',
-        'type': 'numpy',
-    },
-    'occupancy': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/occupancy.zip',
-        'signature': '429cfe376dc9929a1fa528da89f0e1626e34e19695f3f555d8954025bbc522b8',
-        'type': 'numpy',
-    },
-    'mushroom': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/mushroom.zip',
-        'signature': '884c43cb70db35d211c67b1cf6a3683b2b4569393d2789d5c07840da4dc85ba8',
-        'type': 'numpy',
-    },
-    'game': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/game.zip',
-        'signature': 'b1bd85789a014a898daa34cb5f89ceab6d2cd6488a2e572187e34aa4ec21a43b',
-        'type': 'numpy',
-    },
-    'bikeshare': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/bikeshare.zip',
-        'signature': 'a9b440f65549746dff680c92ff8bdca3c7265f09db1cf09e708e6e26fc8aba44',
-        'type': 'numpy',
-    },
-    'spam': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/spam.zip',
-        'signature': '65be21196ba3d8448847409b70a67d761f873f30719c807600eb516d7aef1de1',
-        'type': 'numpy',
-    },
-    'hobbies': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/hobbies.zip',
-        'signature': '415c8f68df1486d5d84a1d1757a5aa3035aef5ad63ede5013c261d622fbd29d8',
-        'type': 'corpus',
-    },
-}
+try:
+    import pandas as pd
+except ImportError:
+    pd = None
 
 
 ##########################################################################
-## General loading utilities
+## Dataset Object
 ##########################################################################
 
-def load_numpy(name, data_home=None, **kwargs):
+class BaseDataset(object):
     """
-    Loads the numpy matrix from the specified data set, downloads it if
-    it hasn't already been downloaded.
+    Base functionality for Dataset and Corpus objects.
     """
 
-    path = find_dataset_path(name, data_home=data_home)
[…]
