### 2018-07-31 e79086e
adds explanation for why we need to check for y_pred array dimensions
