### 2018-07-31 30dbf9e
updates rocauc tests to make sure in the case of a binary decision, macro, micro, and per_class True raises an error. Also removes poof call to reduce image difference error and updates baseline image for test_binary_decision
