### 2018-07-31 059c8b7
flake8 fixes
