### 2018-08-01 db2d903
increased tolerance for AppVeyor and Travis
