### 2018-08-01 10c6614
stringify FilePath objects for Py2
