+++ tests/test_target/test_feature_correlation.py
@@ -1,0 +1,209 @@
+# tests.test_features.test_feature_correlation
+# Test the feature correlation visualizers
+#
+# Author:  Zijie (ZJ) Poh <poh.zijie@gmail.com>
+# Created: Tue Jul 31 20:21:32 2018 -0700
+#
+# Copyright (C) 2018 District Data Labs
+# For license information, see LICENSE.txt
+#
+# ID: test_feature_correlation.py [] poh.zijie@gmail.com $
+
+"""
+Test the feature correlation to dependent variable visualizer.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import sys
+import pytest
+import numpy as np
+import pandas as pd
+import numpy.testing as npt
+import matplotlib.pyplot as plt
+
+from yellowbrick.target import FeatureCorrelation, feature_correlation
+from yellowbrick.exceptions import YellowbrickValueError, YellowbrickWarning
+
+from sklearn import datasets
+
+from tests.base import VisualTestCase
+from tests.dataset import DatasetMixin
+
+
+##########################################################################
+## Feature Correlation Tests
+##########################################################################
+
+class TestFeatureCorrelationVisualizer(VisualTestCase, DatasetMixin):
+    """
+    FeatureCorrelation visualizer
+    """
+
+    data = datasets.load_diabetes()
+    X, y = data['data'], data['target']
+    labels = data['feature_names']
+
+    @pytest.mark.xfail(
+        sys.platform == 'win32', reason="images not close on windows"
+    )
+    def test_feature_correlation_integrated_pearson(self):
+        """
+        Test FeatureCorrelation visualizer with pearson correlation
+        coefficient
+        """
+        viz = FeatureCorrelation()
+        viz.fit(self.X, self.y)
+        viz.poof()
+
+        self.assert_images_similar(viz)
+
+    @pytest.mark.xfail(
+        sys.platform == 'win32', reason="images not close on windows"
+    )
+    def test_feature_correlation_integrated_mutual_info_regression(self):
+        """
+        Test FeatureCorrelation visualizer with mutual information regression
+        """
+        viz = FeatureCorrelation(method='mutual_info')
+        viz.fit(self.X, self.y, random_state=23456)
+        viz.poof()
+
+        self.assert_images_similar(viz)
+
+    @pytest.mark.xfail(
+        sys.platform == 'win32', reason="images not close on windows"
+    )
+    def test_feature_correlation_integrated_mutual_info_classification(self):
+        """
+        Test FeatureCorrelation visualizer with mutual information
+        on wine dataset (classification)
+        """
+        data = datasets.load_wine()
+        X, y = data['data'], data['target']
+
+        viz = FeatureCorrelation(method='mutual_info', classification=True)
+        viz.fit(X, y, random_state=12345)
+        viz.poof()
+
+        self.assert_images_similar(viz)
+
+    def test_feature_correlation_method_not_implemented(self):
+        """
+        Test FeatureCorrelation visualizer with unknown method
+        """
+        method = 'foo'
+        e = ('Method foo not implement; choose from pearson, mutual_info, '
+             'mutual_info_classif')
+        with pytest.raises(YellowbrickValueError, match=e):
+            FeatureCorrelation(method=method)
+
+    def test_feature_correlation_labels_from_index(self):
+        """
+        Test getting feature labels from index
+        """
+        viz = FeatureCorrelation()
+        viz.fit(self.X, self.y)
+
+        npt.assert_array_equal(viz.features_, np.arange(self.X.shape[1]))
+
+    def test_feature_correlation_labels(self):
+        """
+        Test labels as feature labels
+        """
[…]
