### 2018-08-01 b7e9f5b
Sort dict keys for ordering consistency
