### 2018-08-01 7106fd4
updates docs with details about multiclass classification and adds a warning about binary decisions
