### 2018-08-02 5ab40e4
updates docstring examples to finish last requirement of #252
