### 2018-08-02 a5381e1
clarify that game is an object that holds the win lose or draw game dataset
