### 2018-08-02 643c1c1
updates data load functions to be consistant
