+++ docs/api/classifier/rocauc.rst
@@ -14,7 +14,7 @@ This leads to another metric, area under the curve (AUC), which is a computation
     from sklearn.model_selection import train_test_split
 
     # Load the classification data set
-    data = load_data("occupancy")
+    data = load_occupancy()
 
     # Specify the features of interest and the classes of the target
     features = ["temperature", "relative humidity", "light", "C02", "humidity"]
@@ -61,7 +61,7 @@ ROC curves are typically used in binary classification, and in fact the Scikit-L
 .. code::
 
     # Load multi-class classification dataset
-    game = load_game("game")
+    game = load_game()
 
     classes = ["win", "loss", "draw"]
 
