### 2018-08-02 5911c7e
reverts from load_occupancy to load_data in rocauc docs update
