### 2018-08-02 43c2610
pop size and title from kwargs in TSNEVisualizer to bypass internal sklearn TSNE transform pipeline and use in the draw method and adds two little tests just to make sure users can set custom titles or sizes for their TSNEVisualizers to resolve #539
