### 2018-08-02 ab3c59a
use self._title to set the title, if provided
