+++ yellowbrick/text/tsne.py
@@ -2,6 +2,7 @@
 # Implements TSNE visualizations of documents in 2D space.
 #
 # Author:   Benjamin Bengfort <benjamin@bengfort.com>
+# Author:   Rebecca Bilbro <bilbro@gmail.com>
 # Created:  Mon Feb 20 06:33:29 2017 -0500
 #
 # Copyright (C) 2016 Bengfort.com
@@ -166,22 +167,24 @@ class TSNEVisualizer(TextVisualizer):
     NULL_CLASS = None
 
     def __init__(self, ax=None, decompose='svd', decompose_by=50, labels=None,
-               classes=None, colors=None, colormap=None, random_state=None, **kwargs):
-        """
-        Initialize the TSNE visualizer with visual hyperparameters.
-        """
-        super(TSNEVisualizer, self).__init__(ax=ax, **kwargs)
+                 classes=None, colors=None, colormap=None, random_state=None, **kwargs):
 
         # Visual Parameters
         self.labels = labels
         self.colors = colors
         self.colormap = colormap
         self.random_state = random_state
-        self._size = kwargs.pop('size', None)
-        self._title = kwargs.pop('title', None)
 
-        # TSNE Parameters
-        self.transformer_ = self.make_transformer(decompose, decompose_by, kwargs)
+        # Fetch TSNE kwargs from kwargs by popping only keys belonging to TSNE params
+        tsne_kwargs = {
+            key: kwargs.pop(key)
+            for key in TSNE().get_params()
+            if key in kwargs
+        }
+        self.transformer_ = self.make_transformer(decompose, decompose_by, tsne_kwargs)
+
+        # Call super at the end so that size and title are set correctly
+        super(TSNEVisualizer, self).__init__(ax=ax, **kwargs)
 
     def make_transformer(self, decompose='svd', decompose_by=50, tsne_kwargs={}):
         """
@@ -340,14 +343,9 @@ class TSNEVisualizer(TextVisualizer):
         Finalize the drawing by adding a title and legend, and removing the
         axes objects that do not convey information about TNSE.
         """
-        # See if the user has specified a custom title
-        if self._title:
-            self.set_title(self._title)
-        # Otherwise add the default title
-        else:
-            self.set_title(
-                "TSNE Projection of {} Documents".format(self.n_instances_)
-            )
+        self.set_title(
+            "TSNE Projection of {} Documents".format(self.n_instances_)
+        )
 
         # Remove the ticks
         self.ax.set_yticks([])
