### 2018-08-04 b42c711
adds preliminary documentation for the CVScores visualizer, updates the images used in the documentation, updates the docstrings slightly to make sure sphinx renders properly
