### 2018-08-04 615764f
removes unused import from cv score image generation script in docs
