+++ docs/api/model_selection/cv.py
@@ -9,7 +9,6 @@ Generates a CVScores image
 ## Imports
 ##########################################################################
 
-import os
 import pandas as pd
 import matplotlib.pyplot as plt
 
