### 2018-08-05 55c3bb3
Modified per bbengfort comments
