### 2018-08-05 5ae6e32
Fixes TSNEVisualizer title size bug (#542)
In init, fetch TSNE kwargs from kwargs by popping only keys belonging to TSNE params. Also adds test coverage to make sure we reliably get an error if we pass either size or title to the internal sklearn TSNE.
