### 2018-08-05 2cd47f0
Fixes the axes labels in the jointplot image in the docs (#545)
Fixes jointplot image in the documentation and clarifies  documentation of feature visualizers that joint plot is a direct data visualization so that it will be easier to find

Fixes #474
