### 2018-08-06 8e739d1
make it slightly easier for users to directly import ScatterVisualizer now that it's in contrib
