+++ yellowbrick/contrib/__init__.py
@@ -1,0 +1,11 @@
+# yellowbrick.contrib
+#
+# Contains a variety of extra tools and experimental visualizers outside of
+# core support or still in development.
+#
+#
+# ID: __init__.py [] bilbro@gmail.com $
+
+
+
+from .scatter import ScatterViz, ScatterVisualizer, scatterviz
