### 2018-08-06 5a4d421
Adds preliminary documentation for CVScores visualizer (#547)
* encountering some issues when trying to use the size parameter to set the size of the figure on instantiation of the visualizer

* figured out the sizing issue

* coef shape for logistic regression needs to be flattened

* sklearn logistic regression causing issue for feature importances due to difference coef_ shape

* adds preliminary documentation for the CVScores visualizer, updates the images used in the documentation, updates the docstrings slightly to make sure sphinx renders properly

* adds cvscores to the gallery

* remo
[…]
