+++ yellowbrick/model_selection/cv.py
@@ -1,13 +1,11 @@
-
 # coding: utf-8
-
 # In[ ]:
-
 # yellowbrick.model_selection.cv
 #
 #
 # Author:   Prema Damodaran Roman
-
+# Author:   Rebecca Bilbro <bilbro@gmail.com>
+# Updated:  Sat Aug 4 11:52:02 2018 -0500
 #
 # Copyright (C) 2018 District Data Labs
 # For license information, see LICENSE.txt
@@ -30,50 +28,51 @@ from sklearn.model_selection import cross_val_score
 
 class CVScores(ModelVisualizer):
     """
-    CVScores displays cross validation scores as a bar chart and the 
-    average of the scores as a horizontal line
+    CVScores displays cross-validated scores as a bar chart, with the
+    average of the scores plotted as a horizontal line.
 
     Parameters
-    ---------- 
+    ----------
 
     model : a scikit-learn estimator
-            An object that implements ``fit`` and ``predict``, can be a
-            classifier, regressor, or clusterer so long as there is also a valid
-            associated scoring metric.
-            Note that the object is cloned for each validation.
+        An object that implements ``fit`` and ``predict``, can be a
+        classifier, regressor, or clusterer so long as there is also a valid
+        associated scoring metric.
+        Note that the object is cloned for each validation.
 
     ax : matplotlib.Axes object, optional
-            The axes object to plot the figure on.
+        The axes object to plot the figure on.
 
     cv : int, cross-validation generator or an iterable, optional
-            Determines the cross-validation splitting strategy.
-            Possible inputs for cv are:
-                - None, to use the default 3-fold cross-validation,
-                - integer, to specify the number of folds.
-                - An object to be used as a cross-validation generator.
-                - An iterable yielding train/test splits.
+        Determines the cross-validation splitting strategy.
+        Possible inputs for cv are:
 
-            see the scikit-learn
-            `cross-validation guide <http://scikit-learn.org/stable/modules/cross_validation.html>`_
-            for more information on the possible strategies that can be used here.
+        - None, to use the default 3-fold cross-validation,
+        - integer, to specify the number of folds.
+        - An object to be used as a cross-validation generator.
+        - An iterable yielding train/test splits.
+
+        see the scikit-learn
+        `cross-validation guide <http://scikit-learn.org/stable/modules/cross_validation.html>`_
+        for more information on the possible strategies that can be used here.
 
     scoring : string, callable or None, optional, default: None
-                 A string or scorer callable object / function with signature
-                 ``scorer(estimator, X, y)``. 
-    
-                 See scikit-learn `cross-validation guide <http://scikit-learn.org/stable/modules/cross_validation.html>`_
-                 for more information on the possible metrics that can be used.
+        A string or scorer callable object / function with signature
+        ``scorer(estimator, X, y)``.
+
+        See scikit-learn `cross-validation guide <http://scikit-learn.org/stable/modules/cross_validation.html>`_
+        for more information on the possible metrics that can be used.
 
     kwargs : dict
-            Keyword arguments that are passed to the base class and may influence
-            the visualization as defined in other Visualizers.    
-                 
+        Keyword arguments that are passed to the base class and may influence
+        the visualization as defined in other Visualizers.
+
     Examples
     --------
-    
-    >>> from sklearn.model_selection import KFold, cross_val_score, 
-    >>> ShuffleSplit, StratifiedKFold
-    >>> from sklearn import datasets, svm, linear_model 
+
+    >>> from sklearn.model_selection import KFold, cross_val_score
+    >>> from sklearn.model_selection import ShuffleSplit, StratifiedKFold
[…]
