### 2018-08-09 58dc8aa
adds support for user-specified alpha value for scatterplots, which affects the translucency of the plotted points
