### 2018-08-09 d497f24
adds small test to check that user-provided alpha value will modify the scatterplot attribute properly
