### 2018-08-09 09281d8
in quick method for scatterplot, adds alpha param as named kwarg so that it will affect the plot if provided. Also updates the rest of the quick method params to be named kwargs to ensure the quick method test will pass
