### 2018-08-09 a3a1cb7
add mock test for alpha param
