+++ yellowbrick/model_selection/cross_validation.py
@@ -1,0 +1,241 @@
+# yellowbrick.model_selection.cross_validation
+# Implements cross-validation score plotting for model selection.
+#
+# Author:   Prema Damodaran Roman
+# Created:  Wed June 6 2018 13:32:00 -0500
+# Author:   Rebecca Bilbro <bilbro@gmail.com>
+# Updated:  Fri Aug 10 13:15:43 2018 -0500
+#
+# ID: cross_validation.py [7f47800] pdamo24@gmail.com $
+
+"""
+Implements cross-validation score plotting for model selection.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import numpy as np
+import matplotlib.ticker as ticker
+
+from yellowbrick.base import ModelVisualizer
+from sklearn.model_selection import cross_val_score
+
+
+##########################################################################
+## CVScores Visualizer
+##########################################################################
+
+class CVScores(ModelVisualizer):
+    """
+    CVScores displays cross-validated scores as a bar chart, with the
+    average of the scores plotted as a horizontal line.
+
+    Parameters
+    ----------
+
+    model : a scikit-learn estimator
+        An object that implements ``fit`` and ``predict``, can be a
+        classifier, regressor, or clusterer so long as there is also a valid
+        associated scoring metric.
+        Note that the object is cloned for each validation.
+
+    ax : matplotlib.Axes object, optional
+        The axes object to plot the figure on.
+
+    cv : int, cross-validation generator or an iterable, optional
+        Determines the cross-validation splitting strategy.
+        Possible inputs for cv are:
+
+        - None, to use the default 3-fold cross-validation,
+        - integer, to specify the number of folds.
+        - An object to be used as a cross-validation generator.
+        - An iterable yielding train/test splits.
+
+        See the scikit-learn `cross-validation guide <https://goo.gl/FS3VU6>`_
+        for more information on the possible strategies that can be used here.
+
+    scoring : string, callable or None, optional, default: None
+        A string or scorer callable object / function with signature
+        ``scorer(estimator, X, y)``.
+
+        See scikit-learn `cross-validation guide <https://goo.gl/FS3VU6>`_
+        for more information on the possible metrics that can be used.
+
+    kwargs : dict
+        Keyword arguments that are passed to the base class and may influence
+        the visualization as defined in other Visualizers.
+
+    Examples
+    --------
+
+    >>> from sklearn import datasets, svm
+    >>> iris = datasets.load_iris()
+    >>> clf = svm.SVC(kernel='linear', C=1)
+    >>> X = iris.data
+    >>> y = iris.target
+    >>> visualizer = CVScores(model=clf, cv=5, scoring='f1_macro')
+    >>> visualizer.fit(X,y)
+    >>> visualizer.poof()
+
+    Notes
+    -----
+
+    This visualizer is a wrapper for
+    `sklearn.model_selection.cross_val_score <https://goo.gl/4v7dfL>`_.
+
+    Refer to the scikit-learn
+    `cross-validation guide <https://goo.gl/FS3VU6>`_
+    for more details.
+
+    """
+
+    def __init__(self, model, ax=None, cv=None, scoring=None, **kwargs):
+        super(CVScores, self).__init__(model, ax=ax, **kwargs)
+
+        self.cv = cv
+        self.scoring = scoring
+
+    def fit(self, X, y, **kwargs):
+        """
+        Fits the learning curve with the wrapped model to the specified data.
+        Draws training and test score curves and saves the scores to the
+        estimator.
+
+        Parameters
+        ----------
+        X : array-like, shape (n_samples, n_features)
+            Training vector, where n_samples is the number of samples and
+            n_features is the number of features.
+
[…]
