### 2018-08-10 3be9ac4
updates cross validation docs from cv to cross_validation, shows new images with mean score in legend and no adjustment of ylim
