+++ docs/api/model_selection/index.rst
@@ -11,7 +11,7 @@ The currently implemented model selection visualizers are as follows:
 
 -  :doc:`validation_curve`: visualizes how the adjustment of a hyperparameter influences training and test scores to tune the bias/variance trade-off.
 -  :doc:`learning_curve`: shows how the size of training data influences the model to diagnose if a model suffers more from variance error vs. bias error.
--  :doc:`cv`: displays cross-validated scores as a bar chart with average as a horizontal line.
+-  :doc:`cross_validation`: displays cross-validated scores as a bar chart with average as a horizontal line.
 
 Model selection makes heavy use of cross validation to measure the performance of an estimator. Cross validation splits a dataset into a training data set and a test data set; the model is fit on the training data and evaluated on the test data. This helps avoid a common pitfall, overfitting, where the model simply memorizes the training data and does not generalize well to new or unknown input.
 
@@ -22,4 +22,4 @@ There are many ways to define how to split a dataset for cross validation. For m
 
    validation_curve
    learning_curve
-   cv
+   cross_validation
