### 2018-08-10 c50aa9a
adds preliminary tests for CVScores and new baseline images
