### 2018-08-10 fe1e3b1
fix test docstring
