### 2018-08-11 4a54380
increasing the tolerance for the regression test to try to resolve the images not close error
