### 2018-08-11 db4470c
adds prelim page in the docs for teaching yellowbrick
