+++ docs/teaching.rst
@@ -1,0 +1,24 @@
+.. -*- mode: rst -*-
+
+Yellowbrick for Teachers
+========================
+
+For teachers and students of machine learning, Yellowbrick can be used as a framework for teaching and understanding a large variety of algorithms and methods. In fact, Yellowbrick grew out of the teaching data science courses at Georgetown's School of Continuing Studies!
+
+The following slide deck presents an approach to teaching students about the machine learning workflow (the model selection triple), including:
+
+- feature analysis
+- feature importances
+- feature engineering
+- algorithm selection
+- model evaluation for classification and regression
+- cross-validation
+- hyperparameter tuning
+- the scikit-learn API
+
+.. raw:: html
+
+    <iframe src="https://www.slideshare.net/RebeccaBilbro/slideshelf" width="615px" height="470px" frameborder="0" marginwidth="0" marginheight="0" scrolling="no" style="border:none;" allowfullscreen webkitallowfullscreen mozallowfullscreen></iframe>
+
+
+Teachers are welcome to `download the slides <https://www.slideshare.net/RebeccaBilbro/learning-machine-learning-with-yellowbrick>`_ via SlideShare as a PowerPoint deck, and to add them to their course materials to assist in teaching these important concepts.
