### 2018-08-11 410d68f
ups the tolerance for the regression image test to 33
