### 2018-08-11 003210f
adds tests to ensure CVScores correctly passes cv param to internal sklearn classifiers or regressors that have a CV param
