### 2018-08-11 b1189e8
increases the tolerance for the regression image closeness to 36
