### 2018-08-12 7bca813
fixes a few small typos in new YB for teachers documentation
