### 2018-08-12 962c8bb
Finishing up CVscores visualizer (#559)
This PR aims to complete the tasks in #464 to finalize @pdamodaran's new (and most useful!) CVScores visualizer. Here's what's done:

* Modified the legend that labels the average cross-validation score dotted line with the numeric value of self.cv_scores_mean_

* Removed the ylim constraint so that CVScores plots will be more easily comparable across different models.

* Renames the module from cv to cross_validation for readability and accessibility to newcomers.

* Some additional cleanup in the docstrings for nice rendering in the documentation.

* U
[…]
