### 2018-08-13 d873b0a
adds preliminary FAQ page with some of the commonly asked questions and interlinks with the docs index page
