+++ docs/faq.rst
@@ -1,0 +1,162 @@
+.. -*- mode: rst -*-
+
+Frequently Asked Questions
+==========================
+
+Here are some of our frequently asked questions. If you have a question that is not captured here, please feel free to submit it on our `Google Groups Listserv <https://groups.google.com/forum/#!forum/yellowbrick>`_. This is an email list/forum that members of the community can join and respond to each other; you should be able to receive the quickest response here. Please also consider joining the group so you can respond to questions! You can also ask questions on `Stack Overflow <http://stackoverflow.com/questions/tagged/yellowbrick>`_ and tag them with "yellowbrick". Or you can add issues on GitHub. You can also tweet or direct message us on Twitter `@scikit_yb <https://twitter.com/scikit_yb>`_.
+
+
+How can I change the size of a Yellowbrick plot?
+------------------------------------------------
+
+You can change the ``size`` of a plot by passing in the desired dimensions in pixels on instantiation of the visualizer:
+
+.. code:: python
+
+    # Import the visualizer
+    from yellowbrick.features import RadViz
+
+    # Instantiate the visualizer using the ``size`` param
+    visualizer = RadViz(
+        classes=classes, features=features, size=(1080, 720)
+    )
+
+    ...
+
+
+Note: we are considering adding support for passing in ``size`` in inches, but that isn't currently implemented. For a convenient inch-to-pixel converter, check out `www.unitconversion.org <http://www.unitconversion.org/typography/inchs-to-pixels-y-conversion.html>`_.
+
+How can I change the title of a Yellowbrick plot?
+---------------------------------------------------
+
+You can change the ``title`` of a plot by passing in the desired title as a string on instantiation:
+
+
+.. code:: python
+
+    from yellowbrick.classifier import ROCAUC
+    from sklearn.linear_model import RidgeClassifier
+
+    # Create your custom title
+    my_title = "ROCAUC Curves for Multiclass RidgeClassifier"
+
+    # Instantiate the visualizer passing the custom title
+    visualizer = ROCAUC(
+        RidgeClassifier(), classes=classes, title=my_title
+    )
+
+    ...
+
+
+How can I change the color of a Yellowbrick plot?
+-------------------------------------------------
+
+To customize coloring in your plot, use the ``colors`` or ``cmap`` (or ``colormap``) arguments. Note that different visualizers may require slightly different arguments depending on how they construct the plots.
+
+For instance, the :doc:`api/features/manifold` accepts a ``colors`` argument, which for ``discrete`` targets should be the name of one of the :doc:`api/palettes` or a list of `matplotlib colors <https://matplotlib.org/examples/color/named_colors.html>`_ represented as strings:
+For instance, the :doc:`api/features/manifold` accepts a ``colors`` argument, which for ``discrete`` targets should be the name of a palette from the Yellowbrick :doc:`api/palettes` or a list of `matplotlib colors <https://matplotlib.org/examples/color/named_colors.html>`_ represented as strings:
+
+.. code:: python
+
+    from yellowbrick.features.manifold import Manifold
+
+    visualizer = Manifold(
+        manifold="tsne", target="discrete", colors=["teal", "orchid"]
+    )
+
+    ...
+
+
+... whereas for ``continuous`` targets, ``colors`` should be a colormap:
+
+
+.. code:: python
+
+    from yellowbrick.features.manifold import Manifold
+
+    visualizer = Manifold(
+        manifold="isomap", target="continuous", colors="YlOrRd"
+    )
+
+    ...
+
+
+Other visualizers accept a ``cmap`` argument:
+
+.. code:: python
+
[…]
