### 2018-08-15 bcfe40b
Remove unused import
