### 2018-08-15 ed7b007
Remove DatasetMixin
