### 2018-08-19 33aec16
Feature Correlation to Dependent Variable Visualizer (#532)
As seen in Model comparison using a noisy dataset -1, it is useful to compare the pairwise correlation between the features and the dependent variable or target as a bar chart; similar to Rank1D and Rank2D, except that this is for the target only. This new visualizer is part of the target package and gives a similar Rank1D comparison of pearson and mutual information score between the features and the target. 

Fixes #508
