+++ yellowbrick/target/feature_correlation.py
@@ -1,0 +1,320 @@
+# yellowbrick.classifier.feature_correlation
+# Feature correlation to dependent variable visualizer.
+#
+# Author    Zijie (ZJ) Poh <poh.zijie@gmail.com>
+# Created:  Wed Jul 29 15:30:40 2018 -0700
+#
+# ID: feature_correlation.py [] poh.zijie@gmail.com $
+
+"""
+Feature Correlation to Dependent Variable Visualizer.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import numpy as np
+
+from yellowbrick.target.base import TargetVisualizer
+from yellowbrick.utils import is_dataframe
+from yellowbrick.exceptions import YellowbrickValueError, YellowbrickWarning
+
+from sklearn.feature_selection import (mutual_info_classif,
+                                       mutual_info_regression)
+from scipy.stats import pearsonr
+
+
+##########################################################################
+## Class Feature Correlation
+##########################################################################
+
+class FeatureCorrelation(TargetVisualizer):
+    """
+    Displays the correlation between features and dependent variables.
+
+    This visualizer can be used side-by-side with
+    yellowbrick.features.JointPlotVisualizer that plots a feature
+    against the target and shows the distribution of each via a
+    histogram on each axis.
+
+    Parameters
+    ----------
+    ax : matplotlib Axes, default: None
+        The axis to plot the figure on. If None is passed in the current axes
+        will be used (or generated if required).
+
+    method : one of ['pearson', 'mutual_info-regression',
+            'mutual_info-classification'], default: 'pearson'
+        The method to calculate correlation between features and target.
+        'pearson' uses `scipy.stats.pearsonr`.
+        'mutual_info-regression' uses `sklearn.feature_selection.mutual_info_classif`.
+        'mutual_info-classification' uses `sklearn.feature_selection.mutual_info_classif`.
+
+    labels : list, default: None
+        A list of feature names to use. If a DataFrame is passed to fit and
+        features is None, feature names are selected as the column names.
+
+    sort : boolean, default: False
+        If false, the features are are not sorted in the plot; otherwise
+        features are sorted in ascending order of correlation.
+
+    feature_index : list,
+        A list of feature index to include in the plot.
+
+    feature_names : list of feature names
+        A list of feature names to include in the plot.
+        Must have labels or the fitted data is a DataFrame with column names.
+        If feature_index is provided, feature_names will be ignored.
+
+    kwargs : dict
+        Keyword arguments that are passed to the base class and may influence
+        the visualization as defined in other Visualizers.
+
+    Attributes
+    ----------
+    features_ : np.array
+        The feature labels
+
+    scores_ : np.array
+        Correlation between features and dependent variable.
+
+    Examples
+    --------
+
+    >>> viz = FeatureCorrelation()
+    >>> viz.fit(X, y)
+    >>> viz.poof()
+    """
+
+    def _pearsonr(X, y, **kwargs):
+        """
+        Utility function calculate pearson correlation coefficient
+
+        pearsonr returns a correlation coefficient and p-value;
+        currently only fetch the coefficient
+        but can be updated for future implementations that are
+        interested in p-value visualization
+        """
+        return [pearsonr(x, y, **kwargs)[0] for x in np.asarray(X).T]
+
+    correlation_method = {
+        'pearson': _pearsonr,
+        'mutual_info-regression': mutual_info_regression,
+        'mutual_info-classification': mutual_info_classif
+    }
+    correlation_label = {
+        'pearson': 'Pearson Correlation',
+        'mutual_info-regression': 'Mutual Information',
+        'mutual_info-classification': 'Mutual Information'
+    }
+
[…]
