### 2018-08-19 adead40
Adds manual legend drawing utility
Created a module, yellowbrick.draw that is intended to store utility
functions for common matplotlib drawing procedures (e.g. adding a manual
legend or creating a stacked bar chart).

Added manual_legend that can draw a legend with names and colors by
specifying rectangular patches as the artist for the legend. Although
this does not take into account the line style or marker style of the
plot, it does provide an easy way to specify color and label pairs as
the legend to avoid issues like too many labels, or opacity making the
legend unreadable.

Fixes #557
