### 2018-08-19 #864 @arfon
@bbengfort - could you take a look at this list of potential reviewers and identify a few suitable people? https://bit.ly/joss-reviewers
