### 2018-08-19 #864 @bbengfort
@arfon after a quick Python script that scored reviewers based on topics, here's what I've come up with:

1. tacaswell
2. mlgill
3. jkahn
4. jochym
5. jankatins
6. arokem
7. desilinguist
