### 2018-08-19 27fbaef
added colors labels alignment check
