### 2018-08-20 4414a4d
update pcoords, radviz, and residuals to use manual legend
