### 2018-08-20 d54e281
Fixes #491 unicode-decode error on Windows
