### 2018-08-20 ab17f1c
Fix #491 unicode decode issue on Windows
