+++ tests/dataset.py
@@ -17,6 +17,7 @@ Helper functions for tests that utilize downloadable datasets.
 ## Imports
 ##########################################################################
 
+import io
 import os
 import shutil
 import hashlib
@@ -226,7 +227,7 @@ class DatasetMixin(object):
                 files.append(os.path.join(path, cat, name))
                 target.append(cat)
 
-                with open(os.path.join(path, cat, name), 'r', encoding='UTF-8') as f:
+                with io.open(os.path.join(path, cat, name), 'r', encoding='UTF-8') as f:
                     data.append(f.read())
 
         # Return the data bunch for use similar to the newsgroups example
