### 2018-08-21 01e9d4e
adds alpha param to prediction error plot
