+++ yellowbrick/regressor/residuals.py
@@ -91,6 +91,10 @@ class PredictionError(RegressionScoreVisualizer):
     line_color : color
         Defines the color of the best fit line; can be any matplotlib color.
 
+    alpha : float, default: 1.0
+        Specify a transparency where 1 is completely opaque and 0 is completely
+        transparent. This property makes densely clustered points more visible.
+
     kwargs : dict
         Keyword arguments that are passed to the base class and may influence
         the visualization as defined in other Visualizers.
@@ -113,7 +117,7 @@ class PredictionError(RegressionScoreVisualizer):
     """
 
     def __init__(self, model, ax=None, shared_limits=True,
-                 bestfit=True, identity=True, **kwargs):
+                 bestfit=True, identity=True, alpha=1.0, **kwargs):
         # Initialize the visualizer
         super(PredictionError, self).__init__(model, ax=ax, **kwargs)
 
@@ -127,6 +131,7 @@ class PredictionError(RegressionScoreVisualizer):
         self.shared_limits = shared_limits
         self.bestfit = bestfit
         self.identity = identity
+        self.alpha = alpha
 
     def score(self, X, y=None, **kwargs):
         """
@@ -170,7 +175,12 @@ class PredictionError(RegressionScoreVisualizer):
         ax : the axis with the plotted figure
         """
         label = "$R^2 = {:0.3f}$".format(self.score_)
-        self.ax.scatter(y, y_pred, c=self.colors['point'], alpha=0.75, label=label)
+        self.ax.scatter(
+            y,
+            y_pred,
+            c=self.colors['point'],
+            alpha=self.alpha,
+            label=label)
 
         # TODO If score is happening inside a loop, draw would get called multiple times.
         # Ideally we'd want the best fit line to be drawn only once
@@ -240,8 +250,9 @@ class PredictionError(RegressionScoreVisualizer):
         self.ax.legend(loc='best', frameon=True)
 
 
-def prediction_error(model, X, y=None, ax=None, **kwargs):
-    """Quick method:
+def prediction_error(model, X, y=None, ax=None, alpha=1.0, **kwargs):
+    """
+    Quick method:
 
     Plot the actual targets from the dataset against the
     predicted values generated by our model(s).
@@ -287,6 +298,10 @@ def prediction_error(model, X, y=None, ax=None, **kwargs):
     line_color : color
         Defines the color of the best fit line; can be any matplotlib color.
 
+    alpha : float, default: 1.0
+        Specify a transparency where 1 is completely opaque and 0 is completely
+        transparent. This property makes densely clustered points more visible.
+
     kwargs : dict
         Keyword arguments that are passed to the base class and may influence
         the visualization as defined in other Visualizers.
@@ -297,7 +312,7 @@ def prediction_error(model, X, y=None, ax=None, **kwargs):
         Returns the axes that the prediction error plot was drawn on.
     """
     # Instantiate the visualizer
-    visualizer = PredictionError(model, ax, **kwargs)
+    visualizer = PredictionError(model, ax, alpha=alpha, **kwargs)
 
     # Create the train and test splits
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
