### 2018-08-21 31d42df
added legend_labels to address Issue 458
