### 2018-08-21 12fb863
Fixes #491 change corpus loading unicode to ascii
