### 2018-08-21 6473a6e
Fixing unicode issues for 2.7
