### 2018-08-21 493bef3
adds test coverage for alpha param in PredictionError plot
