### 2018-08-21 4f345ab
add alpha param to ResidualsPlot and minimal test coverage
