### 2018-08-21 1bdb756
adds alpha param to TSNEVisualizer and test coverage
