+++ tests/test_text/test_tsne.py
@@ -35,7 +35,11 @@ try:
 except ImportError:
     pandas = None
 
-
+try:
+    from unittest import mock
+except ImportError:
+    import mock
+    
 ##########################################################################
 ## TSNE Tests
 ##########################################################################
@@ -125,7 +129,7 @@ class TestTSNE(VisualTestCase, DatasetMixin):
         tsne = TSNEVisualizer(size=(100, 50))
 
         assert tsne._size == (100, 50)
-        
+
     def test_make_classification_tsne(self):
         """
         Test tSNE integrated visualization on a sklearn classifier dataset
@@ -213,3 +217,27 @@ class TestTSNE(VisualTestCase, DatasetMixin):
 
         tol = 0.1 if six.PY3 else 40
         self.assert_images_similar(tsne, tol=tol)
+
+    def test_alpha_param(self):
+        """
+        Test that the user can supply an alpha param on instantiation
+        """
+        ## produce random data
+        X, y = make_classification(n_samples=200, n_features=100,
+                               n_informative=20, n_redundant=10,
+                               n_classes=3, random_state=42)
+
+        ## Instantiate a TSNEVisualizer, provide custom alpha
+        tsne = TSNEVisualizer(random_state=64, alpha=0.5)
+
+        # Test param gets set correctly
+        assert tsne.alpha == 0.5
+
+        # Mock ax and fit the visualizer
+        tsne.ax = mock.MagicMock(autospec=True)
+        tsne.fit(X, y)
+
+        # Test that alpha was passed to internal matplotlib scatterplot
+        _, scatter_kwargs = tsne.ax.scatter.call_args
+        assert "alpha" in scatter_kwargs
+        assert scatter_kwargs["alpha"] == 0.5
