### 2018-08-21 4be5647
updates tsne baseline images post alpha param addition
