+++ yellowbrick/cluster/icdm.py
@@ -1,0 +1,218 @@
+# yellowbrick.cluster.icdm
+# Implements Intercluster Distance Map visualizations.
+#
+# Author:  Benjamin Bengfort <benjamin@bengfort.com>
+# Created: Tue Aug 21 11:56:53 2018 -0400
+#
+# ID: icdm.py [] benjamin@bengfort.com $
+
+"""
+Implements Intercluster Distance Map visualizations.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import numpy as np
+import matplotlib.pyplot as plt
+
+from sklearn.manifold import MDS, TSNE
+
+from .base import ClusteringScoreVisualizer
+
+from ..utils.helpers import prop_to_size
+from ..exceptions import YellowbrickValueError
+
+from matplotlib.patches import Circle
+
+##########################################################################
+## InterclusterDistance Visualizer
+##########################################################################
+
+class InterclusterDistance(ClusteringScoreVisualizer):
+    """
+    Intercluster distance maps display an embedding of the cluster centers in
+    2 dimensions with the distance to other centers preserved. E.g. the closer
+    to centers are in the visualization, the closer they are in the original
+    feature space. The clusters are sized according to a scoring metric. By
+    default, they are sized by population, e.g. the number of instances that
+    belong to each center. This gives a sense of the relative importance of
+    clusters. Note however, that because two clusters overlap in the 2D space,
+    it does not imply that they overlap in the original feature space.
+
+    Parameters
+    ----------
+
+    Attributes
+    ----------
+    embedded_centers_ : array of shape (n_clusters, 2)
+        The positions of all the cluster centers on the graph.
+
+    scores_ : array of shape (n_clusters,)
+        The scores of each cluster that determine its size on the graph.
+    """
+
+    EMBEDDINGS = {
+        'mds': MDS,
+        'tsne': TSNE,
+    }
+
+    SCORING = {
+        'membership': np.bincount,
+    }
+
+    def __init__(self, model, ax=None, min_size=400, max_size=25000, legend=True,
+                 embedding='mds', scoring='membership', **kwargs):
+        # Initialize the visualizer bases
+        super(InterclusterDistance, self).__init__(model, ax=ax, **kwargs)
+
+        # Ensure that a valid embedding is passed.
+        if embedding.lower() not in self.EMBEDDINGS:
+            raise YellowbrickValueError(
+                "unknown embedding '{}', chose from '{}'".format(
+                    embedding, ", ".join(self.EMBEDDINGS.keys())
+                )
+            )
+
+        # Ensure that a valid scoring metric is passed.
+        if scoring.lower() not in self.SCORING:
+            raise YellowbrickValueError(
+                "unknown scoring '{}', chose from '{}'".format(
+                    scoring, ", ".join(self.SCORING.keys())
+                )
+            )
+
+        # Set decomposition properties
+        self.embedding = embedding
+        self.scoring = scoring
+
+        # Set visual properties
+        self.min_size = min_size
+        self.max_size = max_size
+        self.legend = legend
+
+        # Colors are currently hardcoded, need to compute face and edge color
+        # from this color based on the alpha of the cluster center.
+        self.facecolor = "#2e719344"
+        self.edgecolor = "#2e719399"
+
+    def fit(self, X, y=None):
+        """
+        Fit the clustering model, computing the centers then embeds the centers
+        into 2D space using the embedding method specified.
+        """
+        # Fit the underlying estimator
+        self.estimator.fit(X, y)
+
+        # Get the centers
[…]
