### 2018-08-21 039af95
prop to size py2 correct division
