+++ yellowbrick/utils/helpers.py
@@ -17,6 +17,8 @@ Helper functions and generic utilities for use in Yellowbrick code.
 ## Imports
 ##########################################################################
 
+from __future__ import division
+
 import re
 import numpy as np
 
