### 2018-08-21 84e4f98
bump the tolerance on test_integrated_tsne because the image comparison tests seem to be failing
