### 2018-08-21 5de4eca
If applied this commit denotes document boundaries on DispersionPlot viz
The visualizer was changed so that the expected input is now a list of
documents similiar to hobbies dataset.  This change allows to easily define
the beginning/end of a document and keep track of the word offset
