### 2018-08-21 27d8859
using inset_axes for custom size legend
