+++ yellowbrick/cluster/icdm.py
@@ -17,14 +17,28 @@ Implements Intercluster Distance Map visualizations.
 import numpy as np
 import matplotlib.pyplot as plt
 
+from matplotlib.patches import Circle
 from sklearn.manifold import MDS, TSNE
 
 from .base import ClusteringScoreVisualizer
 
+from ..utils.decorators import memoized
 from ..utils.helpers import prop_to_size
 from ..exceptions import YellowbrickValueError
 
-from matplotlib.patches import Circle
+try:
+    # Only available in Matplotlib >= 2.0.2
+    from mpl_toolkits.axes_grid1 import inset_locator
+except ImportError:
+    inset_locator = None
+
+
+# Valid strings to use for embedding names
+VALID_EMBEDDING = {'mds', 'tsne'}
+
+# Valid strings to use for scoring names
+VALID_SCORING = {'membership',}
+
 
 ##########################################################################
 ## InterclusterDistance Visualizer
@@ -53,49 +67,84 @@ class InterclusterDistance(ClusteringScoreVisualizer):
         The scores of each cluster that determine its size on the graph.
     """
 
-    EMBEDDINGS = {
-        'mds': MDS,
-        'tsne': TSNE,
-    }
 
-    SCORING = {
-        'membership': np.bincount,
-    }
-
-    def __init__(self, model, ax=None, min_size=400, max_size=25000, legend=True,
-                 embedding='mds', scoring='membership', **kwargs):
+    def __init__(self, model, ax=None, min_size=400, max_size=25000,
+                 embedding='mds', scoring='membership',
+                 legend=True, legend_loc="lower left", legend_size=1.5,
+                 random_state=None, **kwargs):
         # Initialize the visualizer bases
         super(InterclusterDistance, self).__init__(model, ax=ax, **kwargs)
 
-        # Ensure that a valid embedding is passed.
-        if embedding.lower() not in self.EMBEDDINGS:
-            raise YellowbrickValueError(
-                "unknown embedding '{}', chose from '{}'".format(
-                    embedding, ", ".join(self.EMBEDDINGS.keys())
-                )
-            )
-
-        # Ensure that a valid scoring metric is passed.
-        if scoring.lower() not in self.SCORING:
-            raise YellowbrickValueError(
-                "unknown scoring '{}', chose from '{}'".format(
-                    scoring, ", ".join(self.SCORING.keys())
-                )
-            )
+        # Ensure that a valid embedding and scoring is passed in
+        validate_embedding(embedding)
+        validate_scoring(scoring)
 
         # Set decomposition properties
         self.embedding = embedding
         self.scoring = scoring
+        self.random_state = random_state
 
         # Set visual properties
         self.min_size = min_size
         self.max_size = max_size
         self.legend = legend
+        self.legend_loc = legend_loc
+        self.legend_size = legend_size
 
         # Colors are currently hardcoded, need to compute face and edge color
-        # from this color based on the alpha of the cluster center.
+        # from this color based on the alpha of the cluster center. The user
+        # can "hack" these properties before drawing, however.
         self.facecolor = "#2e719344"
         self.edgecolor = "#2e719399"
+
+        if self.legend:
+            self.lax # If legend True, test the version availability
+
+    @memoized
+    def lax(self):
+        """
+        Returns the legend axes, creating it only on demand by creating a 2"
+        by 2" inset axes that has no grid, ticks, spines or face frame (e.g
+        is mostly invisible). The legend can then be drawn on this axes.
+        """
+        if inset_locator is None:
+            raise YellowbrickValueError((
+                "intercluster distance map legend requires matplotlib 2.0.2 or greater "
[…]
