+++ docs/api/target/feature_correlation.rst
@@ -6,6 +6,9 @@ Feature Correlation
 This visualizer calculates Pearson correlation coefficients and mutual information between features and the dependent variable.
 This visualization can be used in feature selection to identify features with high correlation or large mutual information with the dependent variable.
 
+Pearson Correlation
+-------------------
+
 The default calculation is Pearson correlation, which is perform with ``scipy.stats.pearsonr``.
 
 .. code:: python
@@ -23,6 +26,9 @@ The default calculation is Pearson correlation, which is perform with ``scipy.st
     visualizer.poof()
 
 .. image:: images/feature_correlation_pearson.png
+
+Mutual Information - Regression
+-------------------------------
 
 Mutual information between features and the dependent variable is calculated with ``sklearn.feature_selection.mutual_info_classif`` when ``method='mutual_info-classification'`` and ``mutual_info_regression`` when ``method='mutual_info-regression'``.
 It is very important to specify discrete features when calculating mutual information because the calculation for continuous and discrete variables are different.
@@ -47,6 +53,9 @@ See [scikit-learn documentation](http://scikit-learn.org/stable/modules/generate
     visualizer.poof()
 
 .. image:: images/feature_correlation_mutual_info_regression.png
+
+Mutual Information - Classification
+-----------------------------------
 
 By fitting with a pandas DataFrame, the feature labels are automatically obtained from the column names.
 This visualizer also allows sorting of the bar plot according to the calculated mutual information (or Pearson correlation coefficients) and selecting features to plot by specifying the names of the features or the feature index.
