### 2018-08-22 e8949d3
changed ax to self.ax
