### 2018-08-22 aa4bbf6
Add clear plt
