### 2018-08-22 c02f54e
finish up tests and docstring, quickmethod
