### 2018-08-22 49d3015
Change set_boundaries parameter to annotate_docs
It is better naming plus confirms to Sklearn metaprogramming
