+++ yellowbrick/text/dispersion.py
@@ -31,49 +31,50 @@ class DispersionPlot(TextVisualizer):
     of words in a corpus.  Lexical dispersion is a measure of a word's
     homeogeneity across the parts of a corpus.  This plot notes the occurences
     of a word and how many words from the beginning it appears.
-    
+
     Parameters
     ----------
     words : list
         A list of target words whose dispersion across a corpus passed at fit
-	will be visualized. 
-    
+	will be visualized.
+
     ax : matplotlib axes, default: None
         The axes to plot the figure on.
-    
+
     color : list or tuple of colors
         Specify color for bars
 
     ignore_case : boolean, default: False
 	Specify whether input  will be case-sensitive.
-        
-    boundary : boolean, default: False
-        Specify whether document boundaries will be drawn
-    
+
+    annotate_docs : boolean, default: False
+        Specify whether document boundaries will be displayed.  Vertical lines
+        are positioned at the end of each document.
+
     kwargs : dict
         Pass any additional keyword arguments to the super class.
-    
+
     These parameters can be influenced later on in the visualization
     process, but can and should be set as early as possible.
     """
-    
+
     def __init__(self, words, ax=None, color=None, ignore_case=False,
-                 boundary=False, **kwargs):
+                 annotate_docs=False, **kwargs):
         super(DispersionPlot, self).__init__(ax=ax, **kwargs)
-        
+
         self.color = color
         self.words = words
         self.ignore_case = ignore_case
-        self.set_boundaries = boundary
-    
+        self.annotate_docs = annotate_docs
+
     def _compute_dispersion(self, text):
         for x, word in enumerate(text):
             if self.ignore_case:
                 word = word.lower()
 
-	    # NOTE: this will find all indices if duplicate words are supplied 
-	    # In the case that word is not in target words, any empty list is
-	    # returned and no data will be yielded 
+            # NOTE: this will find all indices if duplicate words are supplied 
+            # In the case that word is not in target words, any empty list is
+            # returned and no data will be yielded 
             for y in (self.target_words_ == word).nonzero()[0]:
                 yield (x, y)
 
@@ -81,13 +82,13 @@ class DispersionPlot(TextVisualizer):
         """
         The fit method is the primary drawing input for the dispersion
         visualization. It requires the corpus as a list of words.
-        
+
         Parameters
         ----------
         text : list
             A list of words in the order they appear in the corpus.
         """
-            
+
         # Concatenating documents and define boundaries by word offset
         self.boundaries_ = []
         self.doc_words_ = []
@@ -104,12 +105,12 @@ class DispersionPlot(TextVisualizer):
         self.target_words_ = np.flip(self.words, axis=0)
         if self.ignore_case:
             self.target_words_ = np.array([w.lower() for w in self.target_words_])
-        
+
         # Stack is used to create a 2D array from the generator
         points = np.stack(self._compute_dispersion(self.doc_words_))
         self.draw(points)
-        return self 
-    
+        return self
+
     def draw(self, points, **kwargs):
         """
         Called from the fit method, this method creates the canvas and
@@ -118,7 +119,7 @@ class DispersionPlot(TextVisualizer):
         ----------
         kwargs: generic keyword arguments.
         """
-        
+
         # Define boundaries with a solid vertical line
         if self.set_boundaries:
             for xcoords in self.boundaries_:
@@ -127,7 +128,7 @@ class DispersionPlot(TextVisualizer):
         self.ax.scatter(points[:,0], points[:,1], marker='|', color=self.color)
         self.ax.set_yticks(list(range(len(self.target_words_))))
[…]
