### 2018-08-22 c646ac4
documentation for icdm and windows tests
