### 2018-08-23 45bd311
Address memory safety and storage of boundaries values
For memory safety ```the self.boundaries_``` variable needs to be converted to a numpy
array and furthermore, as an extra precaution, the datatype should be set to an integer.

Additionally, we only want to store the boundaries values when the annotate_docs parameter is set.
