### 2018-08-23 ea8eaf5
If applied this commit will add tests for annotating document boundaries on dispersion plots
Changes to Dispersion plot regarding how text is handled by the visualizer requires that we change the
format of the input into `a list of documents containing lists of words` for all the tests.

Additionally, a visual test for annotating_docs parameter was added.
