### 2018-08-24 a1dbc06
Add clear_figure option
