+++ yellowbrick/base.py
@@ -181,7 +181,7 @@ class Visualizer(BaseEstimator):
         """
         return self.ax
 
-    def poof(self, outpath=None, **kwargs):
+    def poof(self, outpath=None, clear_figure=False, **kwargs):
         """
         Poof makes the magic happen and a visualizer appear! You can pass in
         a path to save the figure to disk with various backends, or you can
@@ -209,9 +209,11 @@ class Visualizer(BaseEstimator):
 
         if outpath is not None:
             plt.savefig(outpath, **kwargs)
-            plt.gcf().clear()
         else:
             plt.show()
+
+        if clear_figure:
+            plt.gcf().clear()
 
     ##////////////////////////////////////////////////////////////////////
     ## Helper Functions
@@ -534,7 +536,7 @@ class VisualizerGrid(Visualizer):
 
         return self
 
-    def poof(self, outpath=None, **kwargs):
+    def poof(self, outpath=None, clear_figure=False, **kwargs):
 
         if self.axarr is None: return
 
@@ -553,3 +555,6 @@ class VisualizerGrid(Visualizer):
             plt.savefig(outpath, **kwargs)
         else:
             plt.show()
+
+        if clear_figure:
+            plt.gcf().clear()
