### 2018-08-24 e9ed8db
add fit_time_ to the visualizer
