### 2018-08-24 2ce8d15
Remove itertools chain import
