### 2018-08-24 53ad813
Modify per Kautumn06 comments
