### 2018-08-24 fdbed55
Adds document boundaries option to DispersionPlot (#576)
DispersionPlot has been modified so the expected input is a list of documents
which are in turn a list of words. This change allows us to add an annotate_docs
parameter that draws document boundaries behind the scatter plot (using 
z-order) to show the relative sizes of documents.
