+++ yellowbrick/text/dispersion.py
@@ -31,69 +31,82 @@ class DispersionPlot(TextVisualizer):
     of words in a corpus.  Lexical dispersion is a measure of a word's
     homeogeneity across the parts of a corpus.  This plot notes the occurences
     of a word and how many words from the beginning it appears.
-    
+
     Parameters
     ----------
     words : list
         A list of target words whose dispersion across a corpus passed at fit
-	will be visualized. 
-    
+	will be visualized.
+
     ax : matplotlib axes, default: None
         The axes to plot the figure on.
-    
+
     color : list or tuple of colors
         Specify color for bars
 
     ignore_case : boolean, default: False
 	Specify whether input  will be case-sensitive.
-        
+
+    annotate_docs : boolean, default: False
+        Specify whether document boundaries will be displayed.  Vertical lines
+        are positioned at the end of each document.
+
     kwargs : dict
         Pass any additional keyword arguments to the super class.
-    
+
     These parameters can be influenced later on in the visualization
     process, but can and should be set as early as possible.
     """
-    
-    def __init__(self, words, ax=None, color=None, ignore_case=False, **kwargs):
+
+    def __init__(self, words, ax=None, color=None, ignore_case=False,
+                 annotate_docs=False, **kwargs):
         super(DispersionPlot, self).__init__(ax=ax, **kwargs)
-        
+
         self.color = color
         self.words = words
         self.ignore_case = ignore_case
-        
-    
-    def _compute_dispersion(self, text):
-        for x, word in enumerate(text):
-            if self.ignore_case:
-                word = word.lower()
+        self.annotate_docs = annotate_docs
 
-	    # NOTE: this will find all indices if duplicate words are supplied 
-	    # In the case that word is not in target words, any empty list is
-	    # returned and no data will be yielded 
-            for y in (self.target_words_ == word).nonzero()[0]:
-                yield (x, y)
+    def _compute_dispersion(self, text):
+        self.boundaries_ = []
+        self.offset = 0
+        for doc in text:
+            for word in doc:
+                if self.ignore_case:
+                    word = word.lower()
+
+                # NOTE: this will find all indices if duplicate words are supplied
+                # In the case that word is not in target words, any empty list is
+                # returned and no data will be yielded
+                self.offset += 1
+                for y in (self.target_words_ == word).nonzero()[0]:
+                    yield (self.offset, y)
+            if self.annotate_docs:
+                self.boundaries_.append(self.offset)
+        self.boundaries_ = np.array(self.boundaries_, dtype=int)
 
     def fit(self, text):
         """
         The fit method is the primary drawing input for the dispersion
         visualization. It requires the corpus as a list of words.
-        
+
         Parameters
         ----------
         text : list
-            A list of words in the order they appear in the corpus.
+            Should be provided as a list of documents that contain
+            a list of words in the order they appear in the document.
         """
-            
+
         # Create an index (e.g. the y position) for the target words
         self.target_words_ = np.flip(self.words, axis=0)
         if self.ignore_case:
             self.target_words_ = np.array([w.lower() for w in self.target_words_])
-        
+
         # Stack is used to create a 2D array from the generator
         points = np.stack(self._compute_dispersion(text))
         self.draw(points)
-        return self 
-    
+        return self
+
     def draw(self, points, **kwargs):
         """
         Called from the fit method, this method creates the canvas and
@@ -102,11 +115,17 @@ class DispersionPlot(TextVisualizer):
         ----------
         kwargs: generic keyword arguments.
         """
-        
[…]
