### 2018-08-25 8001bb7
changed self.ax to self
