### 2018-08-26 802ad51
changes some of the default alphas back to 0.7 and adds manual legend to tsne so that low alpha values won't make the legend hard to read
