### 2018-08-26 c9be0b4
no manual legend in prediction error plots for now
