### 2018-08-26 19924a2
adds manual legend import for tsne
