### 2018-08-26 b661023
adds note about why no residuals manual legend for now
