### 2018-08-26 a1bd28b
adds identity line back to PE plot
