### 2018-08-27 d938ce4
adds parameters, attributes, and example to the docstring for SilhouetteVisualizer, fixing #583
