### 2018-08-27 a2acf5a
Establish coloring of markers based on target values
