### 2018-08-27 cb25634
in SilhouetteVisualizer, removes internal properties from init and renames n_samples and n_clusters to n_samples_ and n_clusters_, respectively
