### 2018-08-27 81caf9f
change n_clusters to n_clusters_ in small inline comment
