### 2018-08-27 6037ca3
change default alpha to 0.75 for residuals and prediction error
