### 2018-08-27 d168514
in test_integrated_tsne, manually set alpha to 1.0 and set tolerance back to 40
