### 2018-08-27 cdde432
Finishes SilhouetteVisualizer docstring, completing #583 (#591)
Adds parameters, attributes, and example to SilhouetteVisualizer docstring and removes internal properties from init; fixes #583.
