+++ tests/test_text/test_tsne.py
@@ -85,7 +85,7 @@ class TestTSNE(VisualTestCase, DatasetMixin):
         tsne = TSNEVisualizer(random_state=8392, colormap='Set1', alpha=1.0)
         tsne.fit_transform(docs, labels)
 
-        tol = 40 if six.PY3 else 55
+        tol = 50 if six.PY3 else 55
         self.assert_images_similar(tsne, tol=tol)
 
     def test_sklearn_tsne_size(self):
@@ -178,7 +178,6 @@ class TestTSNE(VisualTestCase, DatasetMixin):
         tsne = TSNEVisualizer(random_state=87, labels=['a', 'b', 'c', 'd'])
         with pytest.raises(YellowbrickValueError):
             tsne.fit(X,y)
-
 
     def test_no_target_tsne(self):
         """
