### 2018-08-27 1e8d330
attempts to fix mocking the axes in alpha scatter_kwargs test for ResidualsPlot test
