### 2018-08-28 41ecf85
Fix indentation in Feature Correlation Visualizer
This commit makes a couple of small edits to the Feature Correlation
Visualizer to make sure that Sphinx renders the documentation correctly.

- Indent two code blocks
- Add double back tick marks to format inline code
- Add a blank line to the docstring following the one-line summary
