### 2018-08-28 b7fc57f
Establish Dispersion Plot with color coded classes including a legend
Fixes labels so that they appear in a legend for the appropriate markers

Problems
--------

I am still faced with the issue of test passing with a generator input and
not target specified.  This results in a ValueError: need at lease one array to stack.
This isn't the case when generator is used with a target specified.  Nor is it the case
when text is input as a list, with or without the target specified.  I have found
that the text is deleted somehow when no target is specified and the generator is used
as input for the t
[…]
