### 2018-08-28 478a336
pull globals out of class and absorbs custom pearson calc into fit method
