+++ yellowbrick/target/feature_correlation.py
@@ -20,10 +20,25 @@ from yellowbrick.target.base import TargetVisualizer
 from yellowbrick.utils import is_dataframe
 from yellowbrick.exceptions import YellowbrickValueError, YellowbrickWarning
 
-from sklearn.feature_selection import (mutual_info_classif,
-                                       mutual_info_regression)
+from sklearn.feature_selection import mutual_info_classif
+from sklearn.feature_selection import mutual_info_regression
+
 from scipy.stats import pearsonr
 
+##########################################################################
+## Supported Correlation Computations
+##########################################################################
+
+CORRELATION_LABELS = {
+    'pearson': 'Pearson Correlation',
+    'mutual_info-regression': 'Mutual Information',
+    'mutual_info-classification': 'Mutual Information'
+}
+
+CORRELATION_METHODS = {
+    'mutual_info-regression': mutual_info_regression,
+    'mutual_info-classification': mutual_info_classif
+}
 
 ##########################################################################
 ## Class Feature Correlation
@@ -43,7 +58,6 @@ class FeatureCorrelation(TargetVisualizer):
     ax : matplotlib Axes, default: None
         The axis to plot the figure on. If None is passed in the current axes
         will be used (or generated if required).
-
 
     method : str, default: 'pearson'
         The method to calculate correlation between features and target.
@@ -91,37 +105,18 @@ class FeatureCorrelation(TargetVisualizer):
     >>> viz.poof()
     """
 
-    def _pearsonr(X, y, **kwargs):
-        """
-        Utility function calculate pearson correlation coefficient
-
-        pearsonr returns a correlation coefficient and p-value;
-        currently only fetch the coefficient
-        but can be updated for future implementations that are
-        interested in p-value visualization
-        """
-        return [pearsonr(x, y, **kwargs)[0] for x in np.asarray(X).T]
-
-        correlation_method = {
-                'pearson': _pearsonr,
-                'mutual_info-regression': mutual_info_regression,
-                'mutual_info-classification': mutual_info_classif
-        }
-        correlation_label = {
-                'pearson': 'Pearson Correlation',
-                'mutual_info-regression': 'Mutual Information',
-                'mutual_info-classification': 'Mutual Information'
-        }
-
     def __init__(self, ax=None, method='pearson',
                  labels=None, sort=False, feature_index=None,
                  feature_names=None, **kwargs):
         super(FeatureCorrelation, self).__init__(ax=None, **kwargs)
 
-        if method not in self.correlation_method:
+        self.correlation_labels = CORRELATION_LABELS
+        self.correlation_methods = CORRELATION_METHODS
+
+        if method not in self.correlation_labels:
             raise YellowbrickValueError(
                 'Method {} not implement; choose from {}'.format(
-                    method, ", ".join(self.correlation_method)
+                    method, ", ".join(self.correlation_labels)
                 )
             )
 
@@ -160,8 +155,13 @@ class FeatureCorrelation(TargetVisualizer):
         self._select_features_to_plot(X)
 
         # Calculate Features correlation with target variable
-        self.scores_ = np.array(
-            self.correlation_method[self.method](X, y, **kwargs)
+        if self.method == "pearson":
+            self.scores_ = np.array(
+                [pearsonr(x, y, **kwargs)[0] for x in np.asarray(X).T]
+            )
+        else:
+            self.scores_ = np.array(
+                self.correlation_methods[self.method](X, y, **kwargs)
         )
 
         # If feature indices are given, plot only the given features
@@ -198,7 +198,7 @@ class FeatureCorrelation(TargetVisualizer):
         """
         self.set_title('Features correlation with dependent variable')
 
[…]
