### 2018-08-28 259b006
[MRG] Add FeatureCorrelation to documentation (#568)
Adds documentation for the new FeatureCorrelation Visualizer. Closes #568
