### 2018-08-29 0b594e5
fix failed test caused by usage of generators as input
Previously, the generator was being exhausted because it was being converted
to a list.  This list was used to count the number of items within the generator
in order to set the exact number of `None` values to set for another operation.
The work around was to use instead `itertools.repeat(None)`

minor
-----
fix docstring
