### 2018-08-29 87ae719
Fix unassigned variable
