### 2018-08-30 b8a6ba6
Removed override of inherited poof method.
