### 2018-09-03 1208ced
Address PR comments
1. Fixed colormap docstring entry by changing wording to "Qualitative colormap for discrete targets"
2. To make it a bit easier to understand, we changed self.words to self.target_words and self.target_words_ to self.indexed_words
3. Changed name of text variable from 'text' to 'X' to make it a bit more similar to how input features are referenced in scikit-learn and in our other text visualizers (e.g. FreqDist and TSNE)
4. Because `self.offset` is only used in `_compute_dispersion` function it is prudent to change it from class attribute to regular variable `offset`
5. Mod
[…]
