### 2018-09-04 62bf834
Make sure that a custom YellowbrickValueError is raised when a target word is not found in the corpus
