### 2018-09-04 b41ce88
Change _check_missing_words loop from enumerate to range(len())
