+++ yellowbrick/classifier/prcurve.py
@@ -1,0 +1,465 @@
+# yellowbrick.classifier.prcurve
+# Implements Precision-Recall curves for classification models.
+#
+# Author:  Benjamin Bengfort <benjamin@bengfort.com>
+# Created: Tue Sep 04 16:47:19 2018 -0400
+#
+# ID: prcurve.py [] benjamin@bengfort.com $
+
+"""
+Implements Precision-Recall curves for classification models.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import numpy as np
+
+from ..exceptions import ModelError, NotFitted
+from ..exceptions import YellowbrickValueError
+from .base import ClassificationScoreVisualizer
+
+from sklearn.preprocessing import label_binarize
+from sklearn.multiclass import OneVsRestClassifier
+from sklearn.utils.multiclass import type_of_target
+from sklearn.metrics import average_precision_score
+from sklearn.model_selection import train_test_split as tts
+from sklearn.metrics import precision_recall_curve as sk_precision_recall_curve
+
+
+# Target Type Constants
+BINARY = "binary"
+MULTICLASS = "multiclass"
+
+# Average Metric Constants
+MICRO = "micro"
+
+
+##########################################################################
+## PrecisionRecallCurve Visualizer
+##########################################################################
+
+class PrecisionRecallCurve(ClassificationScoreVisualizer):
+    """
+    Precision-Recall curves are a metric used to evaluate a classifier's quality,
+    particularly when classes are very imbalanced. The precision-recall curve
+    shows the tradeoff between precision, a measure of result relevancy, and
+    recall, a measure of how many relevant results are returned. A large area
+    under the curve represents both high recall and precision, the best case
+    scenario for a classifier, showing a model that returns accurate results
+    for the majority of classes it selects.
+
+    TODO: extend docstring
+
+    Parameters
+    ----------
+    model : the Scikit-Learn estimator
+        A classification model to score the precision-recall curve on.
+
+    ax : matplotlib Axes, default: None
+        The axes to plot the figure on. If None is passed in the current axes
+        will be used (or generated if required).
+
+    classes : list
+        A list of class names for the legend. If classes is None and a y value
+        is passed to fit then the classes are selected from the target vector.
+        Note that the curves must be computed based on what is in the target
+        vector passed to the ``score()`` method. Class names are used for
+        labeling only and must be in the correct order to prevent confusion.
+
+    fill_area : bool, default=True
+        Fill the area under the curve (or curves) with the curve color.
+
+    ap_score : bool, default=True
+        Annotate the graph with the average precision score, a summary of the
+        plot that is computed as the weighted mean of precisions at each
+        threshold, with the increase in recall from the previous threshold used
+        as the weight.
+
+    micro : bool, default=True
+        If multi-class classification, draw the precision-recall curve for the
+        micro-average of all classes. In the multi-class case, either micro or
+        per-class must be set to True. Ignored in the binary case.
+
+    iso_f1_curves : bool, default=False
+        Draw ISO F1-Curves on the plot to show how close the precision-recall
+        curves are to different F1 scores.
+
+    per_class : bool, default=False
+        If multi-class classification, draw the precision-recall curve for
+        each class using a OneVsRestClassifier to compute the recall on a
+        per-class basis. In the multi-class case, either micro or per-class
+        must be set to True. Ignored in the binary case.
+
+    fill_opacity : float, default=0.2
+        Specify the alpha or opacity of the fill area (0 being transparent,
+        and 1.0 being completly opaque).
+
+    line_opacity : float, default=0.8
[…]
