### 2018-09-05 ea9909a
documentation stubs
