### 2018-09-05 eb5935c
Fixed visual test cases
The fix required more than a simply copy and paste, the tests themselves
had to be modified to ensure that finalize was called. I also resync'd
the baseline images.
