### 2018-09-05 bb1fce5
simplify decision function test
