+++ tests/test_classifier/test_prcurve.py
@@ -110,8 +110,8 @@ class TestPrecisionRecallCurve(VisualTestCase):
         assert 0.0 <= s <= 1.0
 
         # Compare the images
-        # NOTE: do not call finalize to ensure tests pass on CI
-        self.assert_images_similar(oz, tol=1.0)
+        oz.finalize()
+        self.assert_images_similar(oz, tol=3.5)
 
     def test_multiclass_decision(self):
         """
