### 2018-09-05 017b382
implemented predict_proba filter
