### 2018-09-05 588f0e3
increased tolerance for appveyor
