### 2018-09-05 496134e
enhance tests with deeper assertions
