### 2018-09-05 a493062
Add new target module to the documentation's index
On the index.rst page for Yellowbrick's documentation, I created a
target category for the new module. In addition, I moved the
ClassBalance visualizer under it and added the new FeatureCorrelation
visualizer.
