+++ docs/index.rst
@@ -33,7 +33,6 @@ Feature Visualization
 Classification Visualization
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-- :doc:`api/target/class_balance`: see how the distribution of classes affects the model
 - :doc:`api/classifier/class_prediction_error`: shows error and support in classification
 - :doc:`api/classifier/classification_report`: visual representation of precision, recall, and F1
 - :doc:`ROC/AUC Curves <api/classifier/rocauc>`: receiver operator characteristics and area under the curve
@@ -59,6 +58,12 @@ Model Selection Visualization
 
 -  :doc:`api/model_selection/validation_curve`: tune a model with respect to a single hyperparameter
 -  :doc:`api/model_selection/learning_curve`: show if a model might benefit from more data or less complexity
+
+Target Visualization
+~~~~~~~~~~~~~~~~~~~~
+
+- :doc:`api/target/class_balance`: see how the distribution of classes affects the model
+- :doc:`api/target/feature_correlation`: display the correlation between features and dependent variables
 
 Text Visualization
 ~~~~~~~~~~~~~~~~~~
