### 2018-09-06 47c6ef3
Move ClassBalance under Target in the gallery
