### 2018-09-07 5016f16
removes finalize from two jointplot visual comparison tests
