### 2018-09-07 948502e
increases tolerance for jointplot tests
