### 2018-09-10 7796188
Add documentation for BalancedBinningReference
