### 2018-09-10 11eede5
Update index docs with BalancedBinningReference
