### 2018-09-13 ec3e46b
Bugfix to prevent wrong percent points showing up in ConfusionMatrix
