### 2018-09-13 635a24a
adds balanced binning image for docs
