### 2018-09-13 3bb0988
[MRG] Add BalancedBinningReference img to gallery
