### 2018-09-14 b7d6c8a
adds image link under code on binning page in docs
