### 2018-09-14 fe84f81
Add new target module to the documentation's index (#603)
Adds new target module to the documentation's index and adds documentation for Feature Correlation and Balanced Binning visualizers, closes #553 and closes #526.
