### 2018-09-15 fe239b9
responses to PR
