### 2018-09-15 b10d54e
sklearn i, class_i replaces cls
