### 2018-09-15 b40dba1
increase tolerance for appveyor after legend change
