+++ docs/gallery.rst
@@ -128,6 +128,18 @@ Classification Visualizers
     :alt: Multiclass ROC Curves
     :target: api/classifier/rocauc.html#multi-class-rocauc-curves
 
+.. image:: api/classifier/images/binary_precision_recall.png
+    :width: 200px
+    :height: 100px
+    :alt: Precision-Recall Curves
+    :target: api/classifier/prcurve.html
+
+.. image:: api/classifier/images/multiclass_precision_recall_full.png
+    :width: 200px
+    :height: 100px
+    :alt: Multi-Label Precision-Recall Curves
+    :target: api/classifier/prcurve.html#multi-label-classification
+
 .. image:: api/classifier/images/class_prediction_error.png
     :width: 200px
     :height: 100px
