### 2018-09-16 545c6fb
Adds docs for how to use class labels w/ confusion matrix (#611)
Adds documentation about how to use class labels for updated ConfusionMatrix.
