### 2018-09-17 #864 @arfon
@mlgill @jankatins - would you be willing to review this submission for JOSS?
