### 2018-09-17 #864 @jankatins
This looks very interesting! Unfortunately, I'm currently out due to private commitments, sorry!
