### 2018-09-17 #864 @mlgill
@arfon Unfortunately I'm not able to to devote the necessary time to this right now.

Could you let me know how to remove my name from the volunteer list? My work demands have unfortunately shifted dramatically at the moment.
