### 2018-09-17 #864 @arfon
> Could you let me know how to remove my name from the volunteer list? My work demands have unfortunately shifted dramatically at the moment.

Sure thing. I've removed you from the list.
