### 2018-10-01 d750118
Add Kendall-Tau correlation measure
