### 2018-10-02 a4b8450
attempt to jitter manifold test
