### 2018-10-10 3bdb5fb
feat: Add tests to Rank2D
This PR would add tests for testing :
- Pearson,
- Covariance and
- Spearman rank-order correlation

in Rank2D.

Ref: Github issue #628
