### 2018-10-15 4a4f75e
[#609] adding Manifold to __init__.py
