### 2018-10-16 3af6095
Removed changes to example notebook. Fixed and updated k-elbow docs
