### 2018-10-18 8679091
[#609] As per @lwgray suggestion, manifold_embedding is also imported
