### 2018-10-18 e941924
Fix small typos in documentation and docstrings
