### 2018-10-19 60f3726
Fix small typos in documentation and docstrings (#636)
