### 2018-10-23 7133516
changed the scaling of SilhouetteVisualizer's axes
