### 2018-10-23 1a75219
added correct baseline images
