### 2018-10-23 77cdee6
Adding autolayout to rcParams in style
