### 2018-10-25 7664d35
made requested changes and added my test notebook
