### 2018-10-26 7361a8b
Update the text of documentation code examples
I removed the deprecated as_matrix method from the code examples in
Yellowbrick's documentation.
