### 2018-10-28 63716cc
Give full form to the word which easy for understand
