### 2018-10-28 05c8e69
Fix spearman and kendall-tau
