### 2018-10-29 a98847d
Update the text of documentation code examples (#642)
Removed the deprecated as_matrix method from the code examples in Yellowbrick's documentation.
