+++ docs/api/features/radviz.rst
@@ -31,9 +31,9 @@ A good starting place is the `scikit-learn Imputer. <http://scikit-learn.org/sta
     features = ["temperature", "relative humidity", "light", "C02", "humidity"]
     classes = ["unoccupied", "occupied"]
 
-    # Extract the numpy arrays from the data frame
-    X = data[features].as_matrix()
-    y = data.occupancy.as_matrix()
+    # Extract the instances and target
+    X = data[features]
+    y = data.occupancy
 
 .. code:: python
 
