### 2018-10-29 7c8e0fb
Update quickstart.rst
