+++ docs/quickstart.rst
@@ -152,7 +152,9 @@ In this chart, we see that the features ``temp`` and ``feelslike`` have a strong
 
 This visualizer plots a scatter diagram of the apparent temperature on the y axis and the actual measured temperature on the x axis and draws a line of best fit using a simple linear regression. Additionally, univariate distributions are shown as histograms above the x axis for temp and next to the y axis for feelslike.  The ``JointPlotVisualizer`` gives an at-a-glance view of the very strong positive correlation of the features, as well as the range and distribution of each feature. Note that the axes are normalized to the space between zero and one, a common technique in machine learning to reduce the impact of one feature over another.
 
-This plot is very interesting because there appear to be some outliers in the dataset. These instances may need to be manually removed in order to improve the quality of the final model because they may represent data input errors, and potentially train the model on a skewed dataset which would return unreliable model predictions. The first instance of outliers occurs in the temp data where the ``feelslike``  value is approximately equal to 0.25 - showing a horizontal line of data, likely created by input error.
+This plot is very interesting because there appear to be some outliers in the dataset. These instances may need to be manually removed in order to improve the quality of the final model because they may represent data input errors, and potentially train the model on a skewed dataset which would return unreliable model predictions. The first instance of outliers occurs in the temp (this should be "feelslike" column. Because the erroneous inputs are in `feelslike' column. Please check the relevant rows of the dataframe 14130:14156, bike_data.iloc[14130:14156]. Also the parts related with the change shoul change)
+
+data where the ``feelslike``  value is approximately equal to 0.25 - showing a horizontal line of data, likely created by input error.
 
 We can also see that more extreme temperatures create an exaggerated effect in perceived temperature; the colder it is, the colder people are likely to believe it to be, and the warmer it is, the warmer it is perceived to be, with moderate temperatures generally having little effect on individual perception of comfort. This gives us a clue that ``feelslike``  may be a better feature than ``temp`` - promising a more stable dataset, with less risk of running into outliers or errors.
 
