### 2018-10-29 e0415fc
added Iris dataset clustering example and improved scaling of SilhouetteVisualizer (#639)
Improves the scaling of SilhouetteVisualizer's axes to adjust with the data; includes test coverage and example notebooks.
