### 2018-10-29 bad7525
Remove deprecated method from documentation
Removes the deprecated as_matrix method from the Python files used to
generate images in the documentation.

See also: #642
