+++ docs/api/features/pcoords.py
@@ -15,9 +15,9 @@ def load_occupancy_data():
     features = ["temperature", "relative humidity", "light", "C02", "humidity"]
     classes = ['unoccupied', 'occupied']
 
-    # Extract the numpy arrays from the data frame
-    X = data[features].as_matrix()
-    y = data.occupancy.as_matrix()
+    # Extract the instances and target
+    X = data[features]
+    y = data.occupancy
 
     return X, y, features, classes
 
