### 2018-11-02 0c30539
Add tight layout to py file creating Rank2D images
Added plt.tight_layout() to .py file used to create the Rank2D images in
the documentation in order to prevent the labels from being cutoff.
Then, I ran the updated file to regenerate the image .png files for PR #647.
