### 2018-11-14 9657822
Update docs to illustrate changes to the way Dispersion Plot accepts text (#653)
