+++ tests/__init__.py
@@ -28,7 +28,7 @@ matplotlib.use('Agg')
 ## Test Constants
 ##########################################################################
 
-EXPECTED_VERSION = "0.8"
+EXPECTED_VERSION = "0.9"
 
 
 ##########################################################################
