+++ docs/changelog.rst
@@ -3,6 +3,63 @@
 Changelog
 =========
 
+Version 0.9
+-----------
+* Tag: v0.9_
+* Deployed: Wednesday, November 14, 2018
+* Contributors: Rebecca Bilbro, Benjamin Bengfort, Zijie (ZJ) Poh, Kristen McIntyre, Nathan Danielsen, David Waterman, Larry Gray, Prema Roman, Juan Kehoe, Alyssa Batula, Peter Espinosa, Joanne Lin, @rlshuhart, @archaeocharlie, @dschoenleber, Tim Black, @iguk1987, Mohammed Fadhil, Jonathan Lacanlale, Andrew Godbehere, Sivasurya Santhanam, Gopal Krishna
+
+Major Changes:
+    - Target module added for visualizing dependent variable in supervised models.
+    - Prototype missing values visualizer in contrib module.
+    - ``BalancedBinningReference`` visualizer for thresholding unbalanced data (undocumented).
+    - ``CVScores`` visualizer to instrument cross-validation.
+    - ``FeatureCorrelation`` visualizer to compare relationship between a single independent variable and the target.
+    - ``ICDM`` visualizer, intercluster distance mapping using projections similar to those used in pyLDAVis.
+    - ``PrecisionRecallCurve`` visualizer showing the relationship of precision and recall in a threshold-based classifier.
+    - Enhanced ``FeatureImportance`` for multi-target and multi-coefficient models (e.g probabilistic models) and allows stacked bar chart.
+    - Adds option to plot PDF to ``ResidualsPlot`` histogram.
+    - Adds document boundaries option to ``DispersionPlot`` and uses colored markers to depict class.
+    - Added alpha parameter for opacity to the scatter plot visualizer.
+    - Modify ``KElbowVisualizer`` to accept a list of k values.
+    - ``ROCAUC`` bugfix to allow binary classifiers that only have a decision function.
+    - ``TSNE`` bugfix so that title and size params are respected.
+    - ``ConfusionMatrix`` bugfix to correct percentage displays adding to 100.
+    - ``ResidualsPlot`` bugfix to ensure specified colors are both in histogram and scatterplot.
+    - Fixed unicode decode error on Py2 compatible Windows using Hobbies corpus.
+    - Require matplotlib 1.5.1 or matplotlib 2.0 (matplotlib 3.0 not supported yet).
+    - Deprecated percent and sample_weight arguments to ``ConfusionMatrix`` fit method.
+    - Yellowbrick now depends on SciPy 1.0 and scikit-learn 0.20.
+
+Minor Changes:
+    - Removed hardcoding of ``SilhouetteVisualizer`` axes dimensions.
+    - Audit classifiers to ensure they conform to score API.
+    - Fix for ``Manifold`` ``fit_transform`` bug.
+    - Fixed ``Manifold`` import bug.
+    - Started reworking datasets API for easier loading of examples.
+    - Added ``Timer`` utility for keeping track of fit times.
+    - Added slides to documentation for teachers teaching ML/Yellowbrick.
+    - Added an FAQ to the documentation.
+    - Manual legend drawing utility.
+    - New examples notebooks for regression and clustering.
+    - Example of interactive classification visualization using ipywidgets.
+    - Example of using Yellowbrick with PyTorch.
+    - Repairs to ``ROCAUC`` tests and binary/multiclass ``ROCAUC`` construction.
[…]
