+++ requirements.txt
@@ -1,7 +1,7 @@
 ## Dependencies
 matplotlib>=1.5.1,<3.0
-scipy>=0.19
-scikit-learn>=0.19
+scipy>=1.0.0
+scikit-learn>=0.20
 numpy>=1.13.0
 cycler>=0.10.0
 
