### 2018-11-14 47826d9
Added changelog.rst stub for v1.0
New procedure: when we merge PRs we also request that a note is added to the changelog so that we can keep track of all the changes without having to do a full review of all commits on the release day!
