### 2018-11-20 471a0c1
Add test for Kendall-Tau
