### 2018-11-20 a892e83
Update test_rankd.py
