### 2018-11-26 8528a61
Add tests
