### 2018-11-26 1944271
Fix tests
