### 2018-11-26 33a8bdc
Update test_rankd.py
