### 2018-11-30 fd31bd0
Fix Bug mixing colors and labels if y is passed in the fit() and not the constructor of the Visualizer
If the labels are passed only with in the fit() function, there is a mismatch between the colors and the labels due to the fact that the np.unique() method sorts its output.
Do avoid this behavior, one needs to bypass this automatic sorting by working on the unique indexes.
