### 2018-11-30 58c05bd
Update default ranking algorithm in Rank2D docs
The documentation mistakenly says that covariance is the default ranking
algorithm for Rank2D when it is actually Pearson correlation score.

Resolves: #660
