### 2018-11-30 4e38401
Add import statement to Rank2D documentation
In the code example for the Pearson score, it was missing the import statement.

Issue #660
PR #661
