### 2018-12-03 2a414eb
Wrote a UMAPVisualizer class based off the TSNEVisualizer class
