### 2018-12-03 ccd3d2c
cleaned up some of the comments and added a suggestion for the distance function
