### 2018-12-03 9dc4417
added the UMAPVisualizer to the __init__.py for easy class importing
