### 2018-12-03 6657703
Added a series of unit test based off those found in test_tsne.py.  Decomposition test are uneccessary since UMAP has no need of decompositions
