+++ tests/test_text/test_umap.py
@@ -1,0 +1,224 @@
+# tests.test_text.test_umap
+# Tests for the UMAP visual corpus embedding mechanism.
+#
+# Author:   John Healy <jchealy@gmail.com>
+# Created:  Mon Dec 03, 14:00:00
+#
+# Copyright (C) 2016 Bengfort.com
+# For license information, see LICENSE.txt
+#
+#
+
+"""
+Tests for the UMAP visual corpus embedding mechanism.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import six
+import pytest
+
+from yellowbrick.text.umap import *
+from tests.base import VisualTestCase
+from tests.dataset import DatasetMixin
+from yellowbrick.exceptions import YellowbrickValueError
+
+from umap import UMAP
+from sklearn.datasets import make_classification
+from sklearn.feature_extraction.text import TfidfVectorizer
+
+try:
+    import pandas
+except ImportError:
+    pandas = None
+
+try:
+    from unittest import mock
+except ImportError:
+    import mock
+
+##########################################################################
+## TSNE Tests
+##########################################################################
+
+class TestUMAP(VisualTestCase, DatasetMixin):
+    """
+    UMAPVisualizer tests
+    """
+
+    def test_make_pipeline(self):
+        """
+        Verify the pipeline creation step for UMAP
+        """
+
+        umap = UMAPVisualizer() # Should not cause an exception.
+        assert umap.transformer_ is not None
+
+        assert len(umap.transformer_.steps) == 1
+
+    def test_integrated_umap(self):
+        """
+        Check UMAP integrated visualization on the hobbies corpus
+        """
+        corpus = self.load_data('hobbies')
+        tfidf  = TfidfVectorizer()
+
+        docs   = tfidf.fit_transform(corpus.data)
+        labels = corpus.target
+
+        umap = UMAPVisualizer(random_state=8392, colormap='Set1', alpha=1.0)
+        umap.fit_transform(docs, labels)
+
+        tol = 50 if six.PY3 else 55
+        self.assert_images_similar(umap, tol=tol)
+
+    def test_sklearn_umap_size(self):
+        """
+        Check to make sure sklearn's UMAP doesn't use the size param
+        """
+        # In UMAPVisualizer, the internal sklearn UMAP transform consumes
+        # some but not all kwargs passed in by user. Those not in get_params(),
+        # like size, are passed through to YB's finalize method. This test should
+        # notify us  if UMAP's params change on the sklearn side.
+        with pytest.raises(TypeError):
+            UMAP(size=(100,100))
+
+    def test_sklearn_umap_title(self):
+        """
+        Check to make sure sklearn's UMAP doesn't use the title param
+        """
+        # In TSNEVisualizer, the internal sklearn UMAP transform consumes
+        # some but not all kwargs passed in by user. Those not in get_params(),
+        # like title, are passed through to YB's finalize method. This test should
+        # notify us  if UMAP's params change on the sklearn side.
+        with pytest.raises(TypeError):
+            UMAP(title="custom_title")
+
+    def test_custom_title_umap(self):
+        """
+        Check UMAP can accept a custom title (string) from the user
+        """
+        umap = UMAPVisualizer(title="custom_title")
+
+        assert umap.title == "custom_title"
+
+    def test_custom_size_umap(self):
+        """
+        Check UMAP can accept a custom size (tuple of pixels) from the user
+        """
+        umap = UMAPVisualizer(size=(100, 50))
+
+        assert umap._size == (100, 50)
+
+    def test_make_classification_umap(self):
+        """
+        Test UMAP integrated visualization on a sklearn classifier dataset
+        """
+
[…]
