### 2018-12-04 6443275
changed filename to umap_vis.py from umap.py to avoid namespace collision with umap library
