### 2018-12-07 a350608
added the rst and png files for the documentation
