### 2018-12-07 a5c2013
changed to import umap_vis in order to avoid conflict with the default umap package
