### 2018-12-07 fc09634
added umap-learn to the requirements.txt file
