### 2018-12-07 459e382
changed the import in test_umap.py to comply with the new naming scheme of umap_vis
