### 2018-12-07 7cfa229
needed to uncomment development requirements to pass travis tests.
