+++ requirements.txt
@@ -7,27 +7,27 @@ cycler>=0.10.0
 umap-learn>=0.3
 
 ## Testing Requirements (uncomment for development)
-#pytest>=3.4.1
-#pytest-cov>=2.5.1
-#pytest-flakes>=2.0.0
-#pytest-spec>=1.1.0
-#coverage>=4.4.1
-#requests>=2.18.3
-#six==1.11.0
+pytest>=3.4.1
+pytest-cov>=2.5.1
+pytest-flakes>=2.0.0
+pytest-spec>=1.1.0
+coverage>=4.4.1
+requests>=2.18.3
+six==1.11.0
 
 ## Python 2 Testing Requirements
-#mock>=2.0.0
+mock>=2.0.0
 
 ## Optional Testing Dependencies (uncomment for development)
-#nltk>=3.2
-#pandas>=0.20
+nltk>=3.2
+pandas>=0.20
 
 
 ## Documentation (uncomment to build documentation)
-#Sphinx>=1.7.5
-#sphinx-rtd-theme>=0.4.0
-#numpydoc>=0.8.0
+Sphinx>=1.7.5
+sphinx-rtd-theme>=0.4.0
+numpydoc>=0.8.0
 
 ## Build Requirements (uncomment for deployment)
-#wheel>=0.30.0
-#twine>=1.10.0
+wheel>=0.30.0
+twine>=1.10.0
