### 2018-12-10 3c56347
enhance: making nan_counts public properties in bar and dispersion viz
