### 2018-12-10 c880d54
enhance: fix bar doc string.
