+++ yellowbrick/contrib/missing/bar.py
@@ -37,11 +37,9 @@ class MissingValuesBar(MissingDataVisualizer):
 
     Parameters
     ----------
-    alpha : float, default: 0.5
-        A value for bending elments with the background.
 
-    marker : matplotlib marker, default: |
-        The marker used for each element coordinate in the plot
+    width : float, default: 0.5
+        The width of the bars in the visualization.
 
     color : string, default: black
         The color for drawing the bar chart when the y targets are not passed to
