### 2018-12-10 f1d2e86
fix: X and y no longer stored on the viz instance
