### 2018-12-10 0052fa0
enhance: pull out create_nan_matrix into the missing values base class
