### 2018-12-10 fa0a690
enhance: making nan count properties have learned prop convention
