+++ docs/contributing/index.rst
@@ -1,0 +1,39 @@
+.. -*- mode: rst -*-
+
+Contributing
+============
+
+Yellowbrick is an open source project that is supported by a community who will gratefully and humbly accept any contributions you might make to the project. Large or small, any contribution makes a big difference; and if you've never contributed to an open source project before, we hope you will start with Yellowbrick!
+
+Principally, Yellowbrick development is about the addition and creation of *visualizers* --- objects that learn from data and create a visual representation of the data or model. Visualizers integrate with scikit-learn estimators, transformers, and pipelines for specific purposes and as a result, can be simple to build and deploy. The most common contribution is a new visualizer for a specific model or model family. We'll discuss in detail how to build visualizers later.
+
+Beyond creating visualizers, there are many ways to contribute:
+
+- Submit a bug report or feature request on `GitHub issues`_.
+- Contribute an Jupyter notebook to our `examples gallery`_.
+- Assist us with :doc:`user testing <../evaluation>`.
+- Add to the documentation or help with our website, `scikit-yb.org`_
+- Write unit or integration tests for our project.
+- Answer questions on our `GitHub issues`_, `mailing list`_, `Stack Overflow`_, and `Twitter`_.
+- Translate our documentation into another language.
+- Write a blog post, tweet, or share our project with others.
+- Teach someone how to use Yellowbrick.
+
+As you can see, there are lots of ways to get involved and we would be very happy for you to join us! The only thing we ask is that you abide by the principles of openness, respect, and consideration of others as described in our :doc:`../code_of_conduct`.
+
+.. note:: If you're unsure where to start, perhaps the best place is to drop the maintainers a note via our mailing list: http://bit.ly/yb-listserv.
+
+.. _`examples gallery`: https://github.com/DistrictDataLabs/yellowbrick/tree/develop/examples
+.. _`scikit-yb.org`: http://www.scikit-yb.org
+.. _`GitHub issues`: https://github.com/DistrictDataLabs/yellowbrick/issues
+.. _`mailing list`: http://bit.ly/yb-listserv
+.. _`Stack Overflow`: https://stackoverflow.com/questions/tagged/yellowbrick
+.. _`Twitter`: https://twitter.com/scikit_yb
+
+.. toctree::
+    :caption: Contributing Guide
+    :maxdepth: 2
+
+    getting_started
+    developing_visualizers
+    advanced_development_topics
