+++ docs/contributing/advanced_development_topics.rst
@@ -1,0 +1,156 @@
+.. -*- mode: rst -*-
+
+Advanced Development Topics
+===========================
+
+In this section we discuss more advanced contributing guidelines such as code conventions,the release life cycle or branch management. This section is intended for maintainers and core contributors of the Yellowbrick project. If you would like to be a maintainer please contact one of the current maintainers of the project.
+
+Visualizer Review Checklist
+---------------------------
+
+As the visualizer API has matured over time, we've realized that there are a number of routine items that must be in place to consider a visualizer truly complete and ready for prime time. This list is also extremely helpful for reviewing code submissions to ensure that visualizers are consistently implemented, tested, and documented. Though we do not expect these items to be checked off on every PR, the below list includes some guidance about what to look for when reviewing or writing a new Visualizer.
+
+Code Conventions
+~~~~~~~~~~~~~~~~
+
+- Create a quick method for the visualizer.
+
+    In addition to creating the visualizer class, ensure there is an associated quick method that returns the visualizer and creates the visualization in one line of code!
+
+- Subclass the correct visualizer.
+
+    Ensure that the visualizer is correctly subclassed in the class hierarchy. If you're not sure what to subclass, please ping a maintainer, they'd be glad to help!
+
+- Ensure numpy array comparisons are not ambiguous.
+
+    Often there is code such as ``if y:`` where ``y`` is an array. However this is ambiguous when used with numpy arrays and other data containers. Change this code to ``y is not None`` or ``len(y) > 0`` or use ``np.all`` or ``np.any`` to test if the contents of the array are truthy/falsy.
+
+- Add ``random_state`` argument to visualizer.
+
+    If the visualizer uses/wraps a utility that also has ``random_state``, then the visualizer itself needs to also have this argument which defaults to ``None`` and is passed to all internal stochastic behaviors. This ensures that image comparison testing will work and that users can get repeated behavior from visualizers.
+
+- Use ``np.unique`` instead of `set`.
+
+    If you need the unique values from a list or array, we prefer to use numpy methods wherever possible. We performed some limited benchmarking and believe that ``np.unique`` is a bit faster and more efficient.
+
+- Use sklearn underscore suffix for learned parameters.
+
+    Any parameters that are learned during ``fit()`` should only be added to the visualizer when ``fit()`` is called (this is also how we determine if a visualizer is fitted or not) and should be identified with an underscore suffix. For example, in classification visualizers, the classes can be either passed in by the user or determined when they are passed in via fit, therefore it should be ``self.classes_``. This is also true for other learned parameters, e.g. ``self.score_``, even though this is not created during ``fit()``.
+
+Testing Conventions
+~~~~~~~~~~~~~~~~~~~
+
+- Ensure there is an image comparison test.
+
+    Ensure there is at least one image comparison test per visualizer. This is the primary regression testing of Yellowbrick and these tests catch a lot when changes occur in our dependencies or environment.
+
+- Use pytest assertions rather than unittest.
+
+    We prefer ``assert 2+2 == 4`` rather than ``self.assertEquals(2+2, 4)``. Though there is a lot of legacy unittest assertions, we've moved to pytest and one day believe we will have removed the unittest dependency.
+
+- Use test fixtures and sklearn dataset generators.
+
[…]
