+++ requirements.txt
@@ -1,5 +1,5 @@
 ## Dependencies
-matplotlib>=1.5.1,<3.0
+matplotlib>=1.5.1
 scipy>=1.0.0
 scikit-learn>=0.20
 numpy>=1.13.0
