### 2018-12-10 40b27f0
explicitely exclude mpl 3.0.0 due to bug
