+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 # Library Dependencies
-matplotlib>=1.5.1
+matplotlib>=1.5.1,!=3.0.0
 scipy>=0.19
 scikit-learn>=0.19
 numpy>=1.13.0
