### 2018-12-11 9d47f8a
Added code review checklist (#670)
* Added code review checklist

Broke down the contributing.rst file into multiple smaller files for
easier editing. In the advanced development topics section, I added a
section for common code, testing, and documentation conventions, listing
some things that Nathan mentioned and that are part of the ongoing
visualizer audit process.

Fixes #345
