+++ docs/contributing/getting_started.rst
@@ -1,0 +1,93 @@
+.. -*- mode: rst -*-
+
+Getting Started on GitHub
+=========================
+
+Yellowbrick is hosted on GitHub at https://github.com/DistrictDataLabs/yellowbrick.
+
+The typical workflow for a contributor to the codebase is as follows:
+
+1. **Discover** a bug or a feature by using Yellowbrick.
+2. **Discuss** with the core contributors by `adding an issue <https://github.com/DistrictDataLabs/yellowbrick/issues>`_.
+3. **Assign** yourself the task by pulling a card from our `Waffle Kanban <https://waffle.io/DistrictDataLabs/yellowbrick>`_
+4. **Fork** the repository into your own GitHub account.
+5. Create a **Pull Request** first thing to `connect with us <https://github.com/DistrictDataLabs/yellowbrick/pulls>`_ about your task.
+6. **Code** the feature, write the tests and documentation, add your contribution.
+7. **Review** the code with core contributors who will guide you to a high quality submission.
+8. **Merge** your contribution into the Yellowbrick codebase.
+
+.. note:: Please create a pull request as soon as possible, even before you've started coding. This will allow the core contributors to give you advice about where to add your code or utilities and discuss other style choices and implementation details as you go. Don't wait!
+
+We believe that *contribution is collaboration* and therefore emphasize *communication* throughout the open source process. We rely heavily on GitHub's social coding tools to allow us to do this.
+
+Forking the Repository
+----------------------
+
+The first step is to fork the repository into your own account. This will create a copy of the codebase that you can edit and write to. Do so by clicking the **"fork"** button in the upper right corner of the Yellowbrick GitHub page.
+
+Once forked, use the following steps to get your development environment set up on your computer:
+
+1. Clone the repository.
+
+    After clicking the fork button, you should be redirected to the GitHub page of the repository in your user account. You can then clone a copy of the code to your local machine.::
+
+        $ git clone https://github.com/[YOURUSERNAME]/yellowbrick
+        $ cd yellowbrick
+
+2. Create a virtual environment.
+
+    Yellowbrick developers typically use `virtualenv <https://virtualenv.pypa.io/en/stable/>`_ (and `virtualenvwrapper <https://virtualenvwrapper.readthedocs.io/en/latest/>`_), `pyenv <https://github.com/pyenv/pyenv-virtualenv>`_ or `conda envs <https://conda.io/docs/using/envs.html>`_ in order to manage their Python version and dependencies. Using the virtual environment tool of your choice, create one for Yellowbrick. Here's how with virtualenv::
+
+        $ virtualenv venv
+
+3. Install dependencies.
+
+    Yellowbrick's dependencies are in the ``requirements.txt`` document at the root of the repository. Open this file and uncomment the dependencies that are for development only. Then install the dependencies with ``pip``::
+
+        $ pip install -r requirements.txt
+
+    Note that there may be other dependencies required for development and testing; you can simply install them with ``pip``. For example to install
+    the additional dependencies for building the documentation or to run the
+    test suite, use the ``requirements.txt`` files in those directories::
+
+        $ pip install -r tests/requirements.txt
+        $ pip install -r docs/requirements.txt
+
+4. Switch to the develop branch.
+
+    The Yellowbrick repository has a ``develop`` branch that is the primary working branch for contributions. It is probably already the branch you're on, but you can make sure and switch to it as follows::
+
+        $ git fetch
+        $ git checkout develop
+
[…]
