+++ docs/contributing.rst
@@ -1,426 +1,0 @@
-.. -*- mode: rst -*-
-
-Contributing
-============
-
-Yellowbrick is an open source project that is supported by a community who will gratefully and humbly accept any contributions you might make to the project. Large or small, any contribution makes a big difference; and if you've never contributed to an open source project before, we hope you will start with Yellowbrick!
-
-Principally, Yellowbrick development is about the addition and creation of *visualizers* --- objects that learn from data and create a visual representation of the data or model. Visualizers integrate with scikit-learn estimators, transformers, and pipelines for specific purposes and as a result, can be simple to build and deploy. The most common contribution is a new visualizer for a specific model or model family. We'll discuss in detail how to build visualizers later.
-
-Beyond creating visualizers, there are many ways to contribute:
-
-- Submit a bug report or feature request on `GitHub issues`_.
-- Contribute an Jupyter notebook to our `examples gallery`_.
-- Assist us with :doc:`user testing <evaluation>`.
-- Add to the documentation or help with our website, `scikit-yb.org`_
-- Write unit or integration tests for our project.
-- Answer questions on our `GitHub issues`_, `mailing list`_, `Stack Overflow`_, and `Twitter`_.
-- Translate our documentation into another language.
-- Write a blog post, tweet, or share our project with others.
-- Teach someone how to use Yellowbrick.
-
-As you can see, there are lots of ways to get involved and we would be very happy for you to join us! The only thing we ask is that you abide by the principles of openness, respect, and consideration of others as described in our :doc:`code_of_conduct`.
-
-.. note:: If you're unsure where to start, perhaps the best place is to drop the maintainers a note via our mailing list: http://bit.ly/yb-listserv.
-
-.. _`examples gallery`: https://github.com/DistrictDataLabs/yellowbrick/tree/develop/examples
-.. _`scikit-yb.org`: http://www.scikit-yb.org
-.. _`GitHub issues`: https://github.com/DistrictDataLabs/yellowbrick/issues
-.. _`mailing list`: http://bit.ly/yb-listserv
-.. _`Stack Overflow`: https://stackoverflow.com/questions/tagged/yellowbrick
-.. _`Twitter`: https://twitter.com/scikit_yb
-
-Getting Started on GitHub
--------------------------
-
-Yellowbrick is hosted on GitHub at https://github.com/DistrictDataLabs/yellowbrick.
-
-The typical workflow for a contributor to the codebase is as follows:
-
-1. **Discover** a bug or a feature by using Yellowbrick.
-2. **Discuss** with the core contributors by `adding an issue <https://github.com/DistrictDataLabs/yellowbrick/issues>`_.
-3. **Assign** yourself the task by pulling a card from our `Waffle Kanban <https://waffle.io/DistrictDataLabs/yellowbrick>`_
-4. **Fork** the repository into your own GitHub account.
-5. Create a **Pull Request** first thing to `connect with us <https://github.com/DistrictDataLabs/yellowbrick/pulls>`_ about your task.
-6. **Code** the feature, write the tests and documentation, add your contribution.
-7. **Review** the code with core contributors who will guide you to a high quality submission.
-8. **Merge** your contribution into the Yellowbrick codebase.
-
-.. note:: Please create a pull request as soon as possible, even before you've started coding. This will allow the core contributors to give you advice about where to add your code or utilities and discuss other style choices and implementation details as you go. Don't wait!
-
-We believe that *contribution is collaboration* and therefore emphasize *communication* throughout the open source process. We rely heavily on GitHub's social coding tools to allow us to do this.
-
-Forking the Repository
-~~~~~~~~~~~~~~~~~~~~~~
-
[…]
