### 2018-12-11 8695b43
Adds changelog information for recent PRs
