### 2018-12-11 b025d65
Fix subsection headings in documentation
Subsections in the documenation were added using "#" instead of "-"
for the ConfusionMatrix, ROCAUC and CVScores visualizers which caused
the "API Reference" subsection not to be displayed on the Table of
Contents.

I also standardized the headings in the documentation for the
MissingValuesBar and MissingValuesDispersion, even though the "API
Reference" won't be displayed for those visualizers since they are
already nested under "Missing Values" in the TOC.
