### 2018-12-11 62292df
Fix formatting issue in ConfusionMatrix docs
Inline code and variable names in the ConfusionMatrix documentation were
only surrounded by only single backticks, which caused Sphinx to italize the
text, so I added additional backticks where needed to ensure it rendered
properly.

In addition, I added a missing import statement (from sklearn.datasets import load_iris) to the example code.
