### 2018-12-11 143feff
Adds test for DispersionPlot quickmethod
The test for quickmethod for dispersion plot was never
created.  I just overlooked its creation.  This PR
adds an 'assert_images_similar' image comparison for
the quickmethod.
