### 2018-12-11 1ece144
extended the RankD tests and fixed Kendall-Tau tests
