+++ yellowbrick/features/rankd.py
@@ -30,130 +30,26 @@ from yellowbrick.exceptions import YellowbrickValueError
 __all__ = ["rank1d", "rank2d", "Rank1D", "Rank2D"]
 
 
+##########################################################################
+## Metrics
+##########################################################################
+
 def kendalltau(X):
     """
-    Accepts a matrix X and returns a correlation matrix so that each column 
-    is the variable and each row is the observations. 
-    
+    Accepts a matrix X and returns a correlation matrix so that each column
+    is the variable and each row is the observations.
+
     Parameters
     ----------
     X : ndarray or DataFrame of shape n x m
         A matrix of n instances with m features
-    
+
     """
-    corrs = np.zeros(X.shape[1], X.shape[1]) 
+    corrs = np.zeros((X.shape[1], X.shape[1]))
     for idx, cola in enumerate(X.T):
         for jdx, colb in enumerate(X.T):
-            corrs[idx, jdx] = sp_kendalltau(cola,colb)[0] 
-    return corrs 
-
-
-##########################################################################
-## Quick Methods
-##########################################################################
-
-def rank1d(X, y=None, ax=None, algorithm='shapiro', features=None,
-           orient='h', show_feature_names=True, **kwargs):
-    """Scores each feature with the algorithm and ranks them in a bar plot.
-
-    This helper function is a quick wrapper to utilize the Rank1D Visualizer
-    (Transformer) for one-off analysis.
-
-    Parameters
-    ----------
-    X : ndarray or DataFrame of shape n x m
-        A matrix of n instances with m features
-
-    y : ndarray or Series of length n
-        An array or series of target or class values
-
-    ax : matplotlib axes
-        the axis to plot the figure on.
-
-    algorithm : one of {'shapiro', }, default: 'shapiro'
-        The ranking algorithm to use, default is 'Shapiro-Wilk.
-
-    features : list
-        A list of feature names to use.
-        If a DataFrame is passed to fit and features is None, feature
-        names are selected as the columns of the DataFrame.
-
-    orient : 'h' or 'v'
-        Specifies a horizontal or vertical bar chart.
-
-    show_feature_names : boolean, default: True
-        If True, the feature names are used to label the axis ticks in the
-        plot.
-
-    Returns
-    -------
-    ax : matplotlib axes
-        Returns the axes that the parallel coordinates were drawn on.
-
-    """
-    # Instantiate the visualizer
-    visualizer = Rank1D(ax, algorithm, features, orient, show_feature_names,
-                        **kwargs)
-
-    # Fit and transform the visualizer (calls draw)
-    visualizer.fit(X, y, **kwargs)
-    visualizer.transform(X)
-
-    # Return the axes object on the visualizer
-    return visualizer.ax
-
-def rank2d(X, y=None, ax=None, algorithm='pearson', features=None,
-           show_feature_names=True, colormap='RdBu_r', **kwargs):
-    """Displays pairwise comparisons of features with the algorithm and ranks
-    them in a lower-left triangle heatmap plot.
-
-    This helper function is a quick wrapper to utilize the Rank2D Visualizer
-    (Transformer) for one-off analysis.
-
-    Parameters
-    ----------
-    X : ndarray or DataFrame of shape n x m
-        A matrix of n instances with m features
-
-    y : ndarray or Series of length n
-        An array or series of target or class values
-
-    ax : matplotlib axes
-        the axis to plot the figure on.
-
-    algorithm : one of {pearson, covariance, spearman, kendalltau}
-        the ranking algorithm to use, default is Pearson correlation.
-
-    features : list
-        A list of feature names to use.
-        If a DataFrame is passed to fit and features is None, feature
[…]
