+++ tests/test_features/test_rankd.py
@@ -21,13 +21,14 @@ import sys
 import six
 import pytest
 import numpy as np
+import numpy.testing as npt
 
 from tests.base import VisualTestCase
-from tests.dataset import DatasetMixin, Dataset
+from tests.dataset import DatasetMixin
 
 from yellowbrick.features.rankd import *
 from yellowbrick.features.rankd import kendalltau
-from sklearn.datasets import make_classification
+from sklearn.datasets import make_regression
 
 try:
     import pandas as pd
@@ -38,40 +39,65 @@ except ImportError:
 @pytest.fixture(scope='class')
 def dataset(request):
     """
-    Creates a binary classification dataset for use in RankD tests
+    Creates a dataset with 6 gaussian features and 2 categorical features
+    for testing the RankD ranking algorithms. The gaussian features have
+    different correlations with respect to each other, including strong
+    positive and negative correlation and no correlation at all.
     """
-    X, y = make_classification(
-        n_samples=700, n_features=10, n_informative=8, n_redundant=2,
-        n_classes=2, n_clusters_per_class=2, random_state=6483
+    X, _ = make_regression(
+        n_samples=100, n_features=6, effective_rank=2, tail_strength=0,
+        n_informative=2, noise=.45, random_state=27,
     )
 
-    request.cls.dataset = Dataset(X, y)
-    
+    rand = np.random.RandomState(seed=27)
+    request.cls.dataset = np.concatenate((X, rand.binomial(1, 0.6, (100,2))), axis=1)
+
+
 ##########################################################################
 ## Kendall-Tau Tests
 ##########################################################################
 
 @pytest.mark.usefixtures("dataset")
-class TestKendallTau(VisualTestCase, DatasetMixin): 
+class TestKendallTau(object):
     """
-    Test the Kendall-Tau correlation
+    Test the Kendall-Tau correlation metric
     """
-    
-    def test_kendalltau_shape(self):
-        corr = kendalltau(self.dataset.X)
-        self.assertEqual(corr.shape[0], corr.shape[1])
 
-    
+    def test_kendalltau(self):
+        """
+        Test results returned match expectations
+        """
+        expected = np.array([
+            [1.0, -0.68, -0.57454545, 0.49858586, 0.07555556, -0.05858586, 0.02387848, 0.11357219],
+            [-0.68,  1.0, 0.58666667, -0.69090909, -0.22262626, -0.17171717, -0.05059964, -0.12397575],
+            [-0.57454545, 0.58666667, 1.0, -0.61050505, 0.18909091, 0.07515152, 0.00341121, -0.0638663],
+            [0.49858586, -0.69090909, -0.61050505, 1.0, 0.11070707, 0.3030303, 0.03013237, 0.07542581],
+            [ 0.07555556, -0.22262626, 0.18909091, 0.11070707, 1.0, 0.4610101, 0.01648752, 0.05982047],
+            [-0.05858586, -0.17171717, 0.07515152, 0.3030303, 0.4610101, 1.0, 0.03695479, -0.02398599],
+            [0.02387848, -0.05059964, 0.00341121, 0.03013237, 0.01648752, 0.03695479, 1.0, 0.18298883],
+            [0.11357219, -0.12397575, -0.0638663, 0.07542581, 0.05982047, -0.02398599, 0.18298883,  1.0]
+        ])
+        npt.assert_almost_equal(expected, kendalltau(self.dataset))
+
+    def test_kendalltau_shape(self):
+        """
+        Assert that a square correlation matrix is returned
+        """
[…]
