### 2018-12-11 bc542e4
quick method dispersion plot
