### 2018-12-12 e46beda
fixed a few typos in the the rst documentation
