### 2018-12-12 ba9624e
removed conditional on the another image unit test to attempt to pass travis tests.
