### 2018-12-12 8d57fe1
minor changes to pass travis checks
