### 2018-12-14 79b9c0b
numba does not support 32 bit windows python 2.7. Added an error to catch this
