### 2018-12-14 2f76f03
moved umap-learn to an optional dependency section in requirements.txt
