+++ requirements.txt
@@ -4,6 +4,8 @@ scipy>=1.0.0
 scikit-learn>=0.20
 numpy>=1.13.0
 cycler>=0.10.0
+
+## Optional Text Embedding Dependencies (uncomment for umap embedding)
 umap-learn>=0.3
 
 ## Testing Requirements (uncomment for development)
