### 2018-12-14 3e4c4cf
Added runtime error checking to the test_umap.py to catch the fact taht 32 bit windows python 2.7 is not supported by numba.
