### 2018-12-17 15b10f6
added a pytest fail if windows32 since it is not supported on python 2.7
