### 2018-12-17 ca76f2d
trying to pass appveyor tests for windows.
