+++ tests/test_text/test_umap.py
@@ -39,12 +39,12 @@ try:
 except ImportError:
     import mock
 
-pytest.mark.xfail(
+@pytest.mark.xfail(
     sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7"
 )
 
 
-pytest.mark.xfail(
+@pytest.mark.xfail(
     sys.version_info[0] < 3, reason="Having issues with python version 2"
 )
 
