### 2018-12-17 cef13b5
trying to pass appveyor tests for windows.
