### 2018-12-17 3fe2541
trying to pass appveyor tests for windows.
