### 2018-12-17 c421d71
attempt to limit the xfail test to python 2 on win32
