### 2018-12-17 2db9e22
limiting xfail failed.  Rolling back to just checking for win32
