+++ tests/test_text/test_umap.py
@@ -58,7 +58,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
     """
     UMAPVisualizer tests
     """
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_make_pipeline(self):
         """
         Verify the pipeline creation step for UMAP
@@ -69,7 +69,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
 
         assert len(umap.transformer_.steps) == 1
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_integrated_umap(self):
         """
         Check UMAP integrated visualization on the hobbies corpus
@@ -86,7 +86,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
         tol = 55
         self.assert_images_similar(umap, tol=tol)
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_sklearn_umap_size(self):
         """
         Check to make sure sklearn's UMAP doesn't use the size param
@@ -98,7 +98,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
         with pytest.raises(TypeError):
             UMAP(size=(100,100))
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_sklearn_umap_title(self):
         """
         Check to make sure sklearn's UMAP doesn't use the title param
@@ -110,7 +110,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
         with pytest.raises(TypeError):
             UMAP(title="custom_title")
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_custom_title_umap(self):
         """
         Check UMAP can accept a custom title (string) from the user
@@ -119,7 +119,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
 
         assert umap.title == "custom_title"
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_custom_size_umap(self):
         """
         Check UMAP can accept a custom size (tuple of pixels) from the user
@@ -128,7 +128,7 @@ class TestUMAP(VisualTestCase, DatasetMixin):
 
         assert umap._size == (100, 50)
 
-    @pytest.mark.xfail((sys.platform == 'win32') & (sys.version[0] < 3), reason="not supported on windows 32bit with Python 2.7")
+    @pytest.mark.xfail(sys.platform == 'win32', reason="not supported on windows 32bit with Python 2.7")
     def test_make_classification_umap(self):
[…]
