### 2018-12-17 0837120
one more attempt to limit the xfail test to python 2 on win32
