### 2018-12-17 4124b68
one more attempt to limit the xfail test to python 2 on win32
