### 2018-12-22 f5653c6
reincorporates color fix into jerome's tsne PR
