### 2018-12-28 13b534c
Detect continuous vs discrete target for color utils.
Implements a helper function that returns continuous or discrete
depending on the type of target variable, `y`. This function is similar
to the functionality in the Manifold visualizer but makes use of
sklearn.util.multiclass.type_of_target to make its determination, along
with a limit to the number of discrete colors that can be drawn.

Was undecided if this belonged in `yellowbrick.utils` or in
`yellowbrick.target` -- am open to discussion on this topic.

Fixes #73
