### 2018-12-28 5093015
target color type update
