### 2018-12-28 34f5946
minor updates to tests and requirements to finalize PR
