+++ requirements.txt
@@ -5,31 +5,31 @@ scikit-learn>=0.20
 numpy>=1.13.0
 cycler>=0.10.0
 
-## Optional Text Embedding Dependencies (uncomment for umap embedding)
-umap-learn>=0.3
+## Optional Dependencies (uncomment to use)
+#umap-learn>=0.3
 
-## Testing Requirements (uncomment for development)
-pytest>=3.4.1
-pytest-cov>=2.5.1
-pytest-flakes>=2.0.0
-pytest-spec>=1.1.0
-coverage>=4.4.1
-requests>=2.18.3
-six==1.11.0
+## Testing Requirements (pip install -r tests/requirements.txt)
+#pytest>=3.4.1
+#pytest-cov>=2.5.1
+#pytest-flakes>=2.0.0
+#pytest-spec>=1.1.0
+#coverage>=4.4.1
+#requests>=2.18.3
+#six==1.11.0
 
 ## Python 2 Testing Requirements
-mock>=2.0.0
+#mock>=2.0.0
 
-## Optional Testing Dependencies (uncomment for development)
-nltk>=3.2
-pandas>=0.20
+## Optional Testing Dependencies (pip install -r tests/requirements.txt)
+#nltk>=3.2
+#pandas>=0.20
+#umap-learn>=0.3
 
-
-## Documentation (uncomment to build documentation)
-Sphinx>=1.7.5
-sphinx-rtd-theme>=0.4.0
-numpydoc>=0.8.0
+## Documentation (pip install -r docs/requirements.txt)
+#Sphinx>=1.7.5
+#sphinx-rtd-theme>=0.4.0
+#numpydoc>=0.8.0
 
 ## Build Requirements (uncomment for deployment)
-wheel>=0.30.0
-twine>=1.10.0
+#wheel>=0.30.0
+#twine>=1.10.0
