### 2018-12-28 d3a09b1
Updates changelog with recent PRs
Updates v1.0 changelog to reflect new PRs (fixing last name type, dispersion plot quick method, target color type update)
