+++ docs/changelog.rst
@@ -6,17 +6,24 @@ Changelog
 Version 1.0
 -----------
 * Tag: v1.0_
-* Deployed: 
-* Contributors: 
+* Deployed: Not yet deployed
+* Contributors: Benjamin Bengfort, Rebecca Bilbro, Nathan Danielsen, Kristen McIntyre, Larry Gray, Prema Roman, John Healy, Sourav Singh, Francois Dion
 
 Major Changes:
-    - 
-    
+    - Add support for matplotlib 3.0.1 or greater.
+    - ``UMAPVisualizer`` as an alternative manifold to TSNE for corpus visualization that is fast enough to not require preprocessing PCA or SVD decomposition and preserves higher order similarities and distances.
+
 Minor Changes:
-    - 
-    
+    - Updated Rank2D to include Kendall-Tau metric.
+    - Added ``target_color_type`` functionality to determine continuous or discrete color representations based on the type of the target variable.
+    - Added code review checklist and conventions to the documentation and expanded the contributing docs.
+    - Added polish to missing value visualizers code, tests, and documentation.
+    - Improved RankD tests for better coverage.
+    - Added quick method test for ``DispersionPlot`` visualizer.
+    - Improved the documentation readability and structure.
+
 Compatibility Notes:
-    - This version provides support for matplotlib 3.0 and drops support for <matplotlib 2.0 
+    - This version provides support for matplotlib 3.0.1 or greater and drops support for matplotlib versions less than 2.0.
     - This version drops support for Python 2
 
 .. _v1.0: https://github.com/DistrictDataLabs/yellowbrick/releases/tag/v1.0
