### 2018-12-29 e2cd0fd
Bugfix: RFECV step display bug
The RFECV visualizer had a bug when the hyperparameter step > 1. The
step was correctly passed to the internal RFE estimator, which removed
that number of features per iteration, however the feature subsets that
were tried for cross-validation did not match the step resulting in a
figure that looked like no step was actually applied. This patch fixes
the bug and creates a test to ensure this works correctly.

To manage the feature selection subspace, a new learned attribute,
`n_feature_subsets_` was added.

Fixes #664
