### 2018-12-30 91ab4d9
Fix a small typo and remove an extra blank line
* Fix a small typo in the docstring, changing "multilabel-indeicator" to
"multilabel-indicator".

* Remove blank line at the end of `test_target.py`.

PR #676
