### 2018-12-31 7b4e22f
adds test coverage for user-supplied color list in TSNE
