### 2018-12-31 f86810d
fixes analogous resolve colors bug in UMAP visualizer and symmetric test coverage
