### 2018-12-31 1d05fd9
update exceptions
