### 2019-01-01 c316bc2
actually update exceptions
