+++ yellowbrick/datasets/loaders.py
@@ -14,61 +14,54 @@ Dataset loading utilities and primary API to the datasets module.
 ## Imports
 ##########################################################################
 
+import os
+import json
+
 from .base import Dataset, Corpus
+
+__all__ = [
+    "load_concrete", "load_energy", "load_credit", "load_occupancy",
+    "load_mushroom", "load_hobbies", "load_game", "load_bikeshare",
+    "load_spam", "load_walking",
+]
 
 
 ##########################################################################
 ## Links and SHA 256 signature of Yellowbrick hosted datasets
 ##########################################################################
 
-DATASETS = {
-    'concrete': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/concrete.zip',
-        'signature': 'b9ea5f26a7bb272a040e2f1a993b26babbf8dc4a04ab8198bb315ca66d71f10d',
-    },
-    'energy': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/energy.zip',
-        'signature': '19fb86f3bcdde208eed46944172cb643ef6a7d58da103fb568fae43205ed89d3',
-    },
-    'credit': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/credit.zip',
-        'signature': '4a91339c69f55e18f3f48004328fbcb7868070b618208fed099920427b084e5e',
-    },
-    'occupancy': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/occupancy.zip',
-        'signature': '429cfe376dc9929a1fa528da89f0e1626e34e19695f3f555d8954025bbc522b8',
-    },
-    'mushroom': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/mushroom.zip',
-        'signature': '884c43cb70db35d211c67b1cf6a3683b2b4569393d2789d5c07840da4dc85ba8',
-    },
-    'hobbies': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/hobbies.zip',
-        'signature': '415c8f68df1486d5d84a1d1757a5aa3035aef5ad63ede5013c261d622fbd29d8',
-    },
-    'game': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/game.zip',
-        'signature': 'b1bd85789a014a898daa34cb5f89ceab6d2cd6488a2e572187e34aa4ec21a43b',
-    },
-    'bikeshare': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/bikeshare.zip',
-        'signature': 'a9b440f65549746dff680c92ff8bdca3c7265f09db1cf09e708e6e26fc8aba44',
-    },
-    'spam': {
-        'url': 'https://s3.amazonaws.com/ddl-data-lake/yellowbrick/spam.zip',
-        'signature': '65be21196ba3d8448847409b70a67d761f873f30719c807600eb516d7aef1de1',
-    },
-}
+MANIFEST = os.path.join(os.path.dirname(__file__), "manifest.json")
+with open(MANIFEST, "r") as f:
+    DATASETS = json.load(f)
 
 
 ##########################################################################
 ## Specific loading utilities
 ##########################################################################
 
-def load_concrete(data_home=None):
+def _load_dataset(name, data_home=None, return_dataset=False):
+    """
+    Load a dataset by name and return specified format.
+    """
+    info = DATASETS[name]
+    data = Dataset(name, data_home=data_home, **info)
+    if return_dataset:
+        return data
+    return data.to_data()
+
+
+def _load_corpus(name, data_home=None):
+    """
+    Load a corpus object by name.
+    """
+    info = DATASETS[name]
[…]

[…]
