### 2019-01-01 c5e012e
more general .npz access
