### 2019-01-01 b6390b9
py2 fixes for corpus and import error
