### 2019-01-01 5daf006
fixes `resolve_colors` bug in TSNEVisualizer identified by jerome massot (#680)
* fixes resolve colors bug in tsne visualizer identified by jerome massot
* adds test coverage for user-supplied color list in TSNE
* fixes analogous resolve colors bug in UMAP visualizer and symmetric test coverage
