+++ yellowbrick/text/umap_vis.py
@@ -313,7 +313,7 @@ class UMAPVisualizer(TextVisualizer):
 
         # Create the color mapping for the labels.
         self.color_values_ = resolve_colors(
-            n_colors=len(labels), colormap=self.colormap, colors=self.color)
+            n_colors=len(labels), colormap=self.colormap, colors=self.colors)
         colors = dict(zip(labels, self.color_values_))
 
         # Transform labels into a map of class to label
