### 2019-01-01 cad0d25
added PR #680 to the change log
