+++ tests/test_datasets/__init__.py
@@ -1,0 +1,15 @@
+# tests.test_datasets
+# Tests for the datasets module
+#
+# Author:  Benjamin Bengfort <benjamin@bengfort.com>
+# Created: Thu Jul 26 14:28:14 2018 -0400
+#
+# ID: __init__.py [] benjamin@bengfort.com $
+
+"""
+Tests for the datasets module
+"""
+
+##########################################################################
+## Imports
+##########################################################################
