+++ tests/test_datasets/test_path.py
@@ -1,0 +1,230 @@
+# tests.test_datasets.test_paths
+# Tests for the dataset path utilities
+#
+# Author:  Benjamin Bengfort <benjamin@bengfort.com>
+# Created: Thu Jul 26 14:28:14 2018 -0400
+#
+# ID: test_paths.py [] benjamin@bengfort.com $
+
+"""
+Tests for the dataset path utilities
+"""
+
+##########################################################################
+## Imports
+##########################################################################
+
+import os
+import pytest
+import contextlib
+
+from yellowbrick.datasets.path import *
+from yellowbrick.exceptions import DatasetsError
+
+
+##########################################################################
+## Utilities
+##########################################################################
+
+@contextlib.contextmanager
+def environ(**env):
+    """
+    Temporarily set the environment variables for a test, restoring them when
+    the test is complete (e.g. when the context manager exits).
+
+    Parameters
+    ----------
+    env : dict
+        The environment variables that should exist in context
+    """
+    old_env = dict(os.environ)
+    os.environ.clear()
+    os.environ.update(env)
+
+    try:
+        yield
+    finally:
+        os.environ.clear()
+        os.environ.update(old_env)
+
+
+
+##########################################################################
+## Test Cases
+##########################################################################
+
+def test_get_data_home_fixtures():
+    """
+    get_data_home should return fixtures by default
+    """
+    assert get_data_home() == FIXTURES
+    assert os.path.exists(FIXTURES)
+
+
+def test_get_data_home_env(tmpdir):
+    """
+    get_data_home should return the environment variable if set
+    """
+    path = str(tmpdir.mkdir("fixtures").join("foo"))
+    assert not os.path.exists(path)
+
+    with environ(YELLOWBRICK_DATA=path):
+        assert get_data_home() == path
+        assert os.path.exists(path)
+
+
+def test_get_data_home_specified(tmpdir):
+    """
+    get_data_home should return a passed in path
+    """
+    path = str(tmpdir.mkdir("fixtures").join("foo"))
+    assert not os.path.exists(path)
+
+    assert get_data_home(path) == path
+    assert os.path.exists(path)
+
+
+def test_find_dataset_path(tmpdir):
+    """
+    Test find_dataset_path with a specified data_home
+    """
+
+    # Create the dataset
+    data_home = tmpdir.mkdir("fixtures")
+    foo = data_home.mkdir("foo")
+
+    # Stringify FilePath for Py2 tests
+    data_home = str(data_home)
+
+    # Test the default lookup of foo/foo.csv.gz
+    fpath = foo.join("foo.csv.gz")
+    fpath.write("1,2,3")
+    assert find_dataset_path("foo", data_home=data_home) == fpath
+
+    # Test the extension based lookup of foo/foo.npz
+    fpath = foo.join("foo.npz")
+    fpath.write("1234")
+    assert find_dataset_path("foo", data_home=data_home, ext=".npz") == fpath
+
+    # Test the fname based lookup of foo/data.txt
+    fpath = foo.join("data.txt")
+    fpath.write("there is data in this file")
+    assert find_dataset_path("foo", data_home=data_home, fname="data.txt") == fpath
+
+
+def test_missing_find_dataset_path(tmpdir):
+    """
+    Test find_dataset_path when the dataset does not exist
+    """
+    data_home = tmpdir.mkdir("fixtures")
+
+    # When the data directory doesn't exist
+    with pytest.raises(DatasetsError):
[…]
