+++ setup.cfg
@@ -1,5 +1,6 @@
 [metadata]
 description-file = DESCRIPTION.txt
+license_file = LICENSE.txt
 
 [wheel]
 universal = 1
