### 2019-01-08 7f03cb7
Add classes to label binarize y
