+++ yellowbrick/classifier/prcurve.py
@@ -173,7 +173,8 @@ class PrecisionRecallCurve(ClassificationScoreVisualizer):
             self.estimator = OneVsRestClassifier(self.estimator)
 
             # Use label_binarize to create multi-label ouptut for OneVsRestClassifier
-            Y = label_binarize(y, classes=np.unique(y))
+            self.classes_from_y = np.unique(y)
+            Y = label_binarize(y, classes=self.classes_from_y)
         elif ttype.startswith(BINARY):
             self.target_type_ = BINARY
 
@@ -219,7 +220,7 @@ class PrecisionRecallCurve(ClassificationScoreVisualizer):
             self.score_ = average_precision_score(y, y_scores)
         else:
             # Use label_binarize to create multi-label ouptut for OneVsRestClassifier
-            Y = label_binarize(y, classes=self.classes_)
+            Y = label_binarize(y, classes=self.classes_from_y)
 
             self.precision_, self.recall_, self.score_ = {}, {}, {}
 
