### 2019-01-14 186cc6e
marking out the script I will be adding my changes to in the future
