### 2019-01-22 5622f00
enhance: add installing test suite dependencing to contrib section with some small rewrites
enhance: add link in readme to testing instructions

add installation section
