+++ tests/test_gridsearch/__init__.py
@@ -1,0 +1,11 @@
+# tests.test_gridsearch
+# Tests for the Grid search visualizer library.
+#
+
+"""
+Tests for the model selection visualizer library.
+"""
+
+##########################################################################
+## Imports
+##########################################################################
