### 2019-01-24 c823d2f
Fix code to autocreate elbow plot images in docs
In order to use matplotlib's sphinx extension (PR #446), each code block in the
documentation used to generate an image must be complete. For instance,
this commit updates the docs for Yellowbrick's KElbowVisualizer, which
previously only used sklearn's make_blobs function in the first code
block to create the synthetic dataset used for both of our examples. Therefore, I needed to add the make_blobs function to the second block of code, since even though the code blocks are on the same page, they are entirely separate.

See also: #354, #353
