### 2019-01-28 9cd8ad0
added feature to specify custom f1 values for iso_f1_curve and added feature to specify train test data in pr quick method
