+++ yellowbrick/classifier/prcurve.py
@@ -5,7 +5,6 @@
 # Created: Tue Sep 04 16:47:19 2018 -0400
 #
 # ID: prcurve.py [] benjamin@bengfort.com $
-# TODO : Need to work on this script to address issue #610
 """
 Implements Precision-Recall curves for classification models.
 """
