### 2019-01-30 3de4729
fix structured arrays
