### 2019-01-30 a4aa03a
fix scatter viz image test
