### 2019-01-30 cf0fab0
bump py2 tolerance for Travis
