### 2019-01-30 3a79071
xfail confusion matrix quick method test on windows
