### 2019-01-30 870022a
Fixes CI Tests
Repairs breaking tests to resolve travis pyflakes error and appveyer value error, updates baseline images, adds skips for some unresolved tests from contrib package.
