### 2019-01-31 5e237ce
referenced to the links
I changed the jointplots and rank2d links,so that they can directly go to their respective docs pages
