+++ docs/quickstart.rst
@@ -122,9 +122,7 @@ After downloading the dataset and unzipping it in your current working directory
     ]]
     y = data["riders"]
 
-The machine learning workflow is the art of creating *model selection triples*, a combination of features, algorithm, and hyperparameters that uniquely identifies a model fitted on a specific data set. As part of our feature selection, we want to identify features that have a linear relationship with each other, potentially introducing covariance into our model and breaking OLS (guiding us toward removing features or using regularization). We can use the Rank2D_ visualizer to compute Pearson correlations between all pairs of features as follows:
-
-.. _Rank2D: https://github.com/DistrictDataLabs/yellowbrick/blob/develop/docs/index.rst#id13
+The machine learning workflow is the art of creating *model selection triples*, a combination of features, algorithm, and hyperparameters that uniquely identifies a model fitted on a specific data set. As part of our feature selection, we want to identify features that have a linear relationship with each other, potentially introducing covariance into our model and breaking OLS (guiding us toward removing features or using regularization). We can use the :ref:`Rank2D <:doc:`Rank2D <api/features/rankd>`:Rank2D>` visualizer to compute Pearson correlations between all pairs of features as follows:
 
 .. code-block:: python
 
@@ -138,7 +136,7 @@ The machine learning workflow is the art of creating *model selection triples*, 
 
 This figure shows us the Pearson correlation between pairs of features such that each cell in the grid represents two features identified in order on the x and y axes and whose color displays the magnitude of the correlation. A Pearson correlation of 1.0 means that there is a strong positive, linear relationship between the pairs of variables and a value of -1.0 indicates a strong negative, linear relationship (a value of zero indicates no relationship). Therefore we are looking for dark red and dark blue boxes to identify further.
 
-In this chart, we see that the features ``temp`` and ``feelslike`` have a strong correlation and also that the feature ``season`` has a strong correlation with the feature ``month``. This seems to make sense; the apparent temperature we feel outside depends on the actual temperature and other airquality factors, and the season of the year is described by the month! To dive in deeper, we can use the `JointPlotVisualizer https://github.com/DistrictDataLabs/yellowbrick/blob/develop/docs/index.rst#id27`_ to inspect those relationships.
+In this chart, we see that the features ``temp`` and ``feelslike`` have a strong correlation and also that the feature ``season`` has a strong correlation with the feature ``month``. This seems to make sense; the apparent temperature we feel outside depends on the actual temperature and other airquality factors, and the season of the year is described by the month! To dive in deeper, we can use the :doc:`Joint Plots <api/features/jointplot>` to inspect those relationships.
 
 .. code-block:: python
 
