### 2019-01-31 a2f9b0a
reorg of doc configuration and requirements
