### 2019-01-31 b30964b
fixed links
Rank2D and JointPlots links fixed
