+++ tests/conftest.py
@@ -25,6 +25,17 @@ from pytest_flakes import FlakesItem
 ## PyTest Hooks
 ##########################################################################
 
+def docline(obj):
+    """
+    Returns the first line of the object's docstring or None if
+    there is no __doc__ on the object.
+    """
+    if not obj.__doc__:
+        return None
+    lines = list(filter(None, obj.__doc__.split("\n")))
+    return lines[0].strip()
+
+
 def pytest_itemcollected(item):
     """
     A reporting hook that is called when a test item is collected.
@@ -49,11 +60,11 @@ def pytest_itemcollected(item):
     # or class name, and the docstring of the test case, then set the nodeid
     # so that pytest-spec will correctly parse the information.
     path = os.path.relpath(str(item.fspath))
-    prefix = parent.__doc__ or getattr(parent, '__name__', parent.__class__.__name__)
-    suffix = node.__doc__.strip() if node.__doc__ else node.__name__
+    prefix = docline(parent) or getattr(parent, '__name__', parent.__class__.__name__)
+    suffix = docline(node) or node.__name__
 
     # Add parametrize or test generation id to distinguish it in output
-    if item._genid:
+    if hasattr(item, "_genid") and item._genid:
         suffix += " ({})".format(item._genid)
 
     if prefix or suffix:
