### 2019-01-31 eb5d03f
remove spec while warnings are being issued
