### 2019-02-01 b75f3bd
Documents the yellowbrick.download script.
Adds documentation for the `yellowbrick.download` script that is
included for dataset management. The documentation is located in both
the README.md and in the contributor's guide. This documentation will
hopefully also assist developers who are having trouble with older
dataset versions on their computers.

Closes #693
