### 2019-02-01 05e0769
Auto-generate Images for Documentation
POC for auto-generation of images in the scikit-yb docs, focused on feature and cluster visualizers.
