### 2019-02-01 948ea95
Repairs broken links for rank2d and jointplots
@rebeccabilbro
