### 2019-02-01 ea61c5a
corrections to rst for walkthrough
