### 2019-02-02 47b4ee4
Create effect.py
