### 2019-02-02 02e62c0
Adds plot directives to quickstart
This PR was primarily intended to add plot directives to the quickstart
guide, ensuring that the images in the tutorial were always up to date
with the library. The quickstart guide separates the code blocks from
the plot directives so that the code is a linear narrative rather than
the verbosity required for each independent code block. This duplicates
the code a little bit but makes it more readable.

To ensure that the code is correct I also created a notebook
examples/walkthrough.ipynb with the code from the quickstart.

Along the way there were some bugs
[…]
