### 2019-02-02 270caac
streamlines readme, replaces flickr image with gallery of visualizers, includes script to generate readme images
