+++ README.md
@@ -13,70 +13,15 @@
 
 **Visual analysis and diagnostic tools to facilitate machine learning model selection.**
 
-![Follow the yellow brick road](docs/images/yellowbrickroad.jpg)
-Image by [Quatro Cinco](https://flic.kr/p/2Yj9mj), used with permission, Flickr Creative Commons.
-
-This README is a guide for developers, if you're new to Yellowbrick, get started at our [documentation](http://www.scikit-yb.org/).
-
 ## What is Yellowbrick?
 
-Yellowbrick is a suite of visual diagnostic tools called "Visualizers" that extend the scikit-learn API to allow human steering of the model selection process. In a nutshell, Yellowbrick combines scikit-learn with matplotlib in the best tradition of the scikit-learn documentation, but to produce visualizations for _your_ models!
+Yellowbrick is a suite of visual diagnostic tools called "Visualizers" that extend the scikit-learn API to allow human steering of the model selection process. In a nutshell, Yellowbrick combines scikit-learn with matplotlib in the best tradition of the scikit-learn documentation, but to produce visualizations for _your_ machine learning workflow!
 
-![Visualizers](docs/images/visualizers.png)
+![Classifiers](docs/images/readme_classifiers.png)
+![Clusterers](docs/images/readme_clusterers.png)
+![Regressors](docs/images/readme_regressors.png)
 
-### Visualizers
-
-Visualizers are estimators (objects that learn from data) whose primary objective is to create visualizations that allow insight into the model selection process. In scikit-learn terms, they can be similar to transformers when visualizing the data space or wrapping a model estimator similar to how the "ModelCV" (e.g. RidgeCV, LassoCV) methods work. The primary goal of Yellowbrick is to create a sensical API similar to scikit-learn. Some of our most popular visualizers include:
-
-#### Feature Visualization
-
-- **Rank Features**: single or pairwise ranking of features to detect relationships
-- **Parallel Coordinates**: horizontal visualization of instances
-- **Radial Visualization**: separation of instances around a circular plot
-- **PCA Projection**: projection of instances based on principal components
-- **Manifold Visualization**: high dimensional visualization with manifold learning
-- **Feature Importances**: rank features based on their in-model performance
-- **Recursive Feature Elimination**: find the best subset of features by importance
-- **Joint Plots**: direct data visualization with feature selection
-
-#### Classification Visualization
-
-- **Class Balance**: see how the distribution of classes affects the model
-- **Class Prediction Error**: shows error and support in classification
-- **Classification Report**: visual representation of precision, recall, and F1
-- **ROC/AUC Curves**: receiver operator characteristics and area under the curve
-- **Precision-Recall Curves**: precision vs recall for different probability thresholds
-- **Confusion Matrices**: visual description of class decision making
-- **Discrimination Threshold**: find a threshold that best separates binary classes
-
-#### Regression Visualization
-
-- **Prediction Error Plots**: find model breakdowns along the domain of the target
-- **Residuals Plot**: show the difference in residuals of training and test data
-- **Alpha Selection**: show how the choice of alpha influences regularization
-
-#### Clustering Visualization
-
-- **K-Elbow Plot**: select k using the elbow method and various metrics
-- **Silhouette Plot**: select k by visualizing silhouette coefficient values
-- **Intercluster Distance Maps**: show relative distance and size of clusters
-
-#### Model Selection Visualization
-
-- **Validation Curve**: tune a model with respect to a single hyperparameter
-- **Learning Curve**: show if a model might benefit from more data or less complexity
-
-#### Text Visualization
-
-- **Term Frequency**: visualize the frequency distribution of terms in the corpus
-- **t-SNE Corpus Visualization**: use stochastic neighbor embedding to project documents.
[…]
