### 2019-02-02 8abb65f
move gallery images higher, add links to docs gallery
