### 2019-02-03 e8fe0ac
Refresh the JointPlot Visualizer for ML
Our original (and wonderful, thank you @pdamodaran) JointPlotVisualizer
was based off of seaborn's implementation and used mpl.GridSpec for its
layout. Since that implementation we've grown a bit more sophisticated
in our axes management, but JointPlot hasn't been updated to reflect
this.

This PR refreshes JointPlot with the make_axes_locatable method of
layout, similar to what we use in the ResidualsPlot. Additionally this
PR rethinks the JointPlot from the ML perspective of passing data as X
and y, rather than seaborn's more data-standard approach. Fo
[…]
