+++ yellowbrick/features/jointplot.py
@@ -1,11 +1,11 @@
 
 # yellowbrick.features.jointplot
-# Implementations of joint plots for univariate and bivariate analysis.
+# Implementation of joint plots for univariate and bivariate analysis.
 #
 # Author:   Prema Damodaran Roman
 # Created:  Mon Apr 10 21:00:54 2017 -0400
 #
-# Copyright (C) 2017 District Data Labs
+# Copyright (C) 2017 The scikit-yb developers.
 # For license information, see LICENSE.txt
 #
 # ID: jointplot.py [7f47800] pdamodaran@users.noreply.github.com $
@@ -14,19 +14,30 @@
 ## Imports
 ##########################################################################
 
-import warnings
 import numpy as np
-import matplotlib as mpl
 import matplotlib.pyplot as plt
 
-from yellowbrick.features.base import FeatureVisualizer
-from yellowbrick.exceptions import YellowbrickValueError
-from yellowbrick.bestfit import draw_best_fit
-from yellowbrick.utils import is_dataframe
+try:
+    # Only available in Matplotlib >= 2.0.2
+    from mpl_toolkits.axes_grid1 import make_axes_locatable
+except ImportError:
+    make_axes_locatable = None
+
+from scipy.stats import pearsonr
+from .base import FeatureVisualizer
+# from ..bestfit import draw_best_fit
+from ..exceptions import YellowbrickValueError
 
 
+# Default Colors
 FACECOLOR = "#FAFAFA"
 HISTCOLOR = "#6897bb"
+
+
+# Objects for export
+__all__ = [
+    "JointPlot", "JointPlotVisualizer", "joint_plot",
+]
 
 
 ##########################################################################
@@ -34,288 +45,338 @@ HISTCOLOR = "#6897bb"
 ##########################################################################
 
 
-class JointPlotVisualizer(FeatureVisualizer):
+class JointPlot(FeatureVisualizer):
     """
-    JointPlotVisualizer allows for a simultaneous visualization of the relationship
-    between two variables and the distrbution of each individual variable.  The
-    relationship is plotted along the joint axis and univariate distributions
-    are plotted on top of the x axis and to the right of the y axis.
 
     Parameters
     ----------
-    ax: matplotlib Axes, default: None
-        This is inherited from FeatureVisualizer but is defined within
-        JointPlotVisualizer since there are three axes objects.
+    ax : matplotlib Axes, default: None
+        The axes to plot the figure on. If None is passed in the current axes will be
+        used (or generated if required). This is considered the base axes where the
+        the primary joint plot is drawn. It will be shifted and two additional axes
+        added above (xhax) and to the right (yhax) if hist=True.
 
-    feature: string, default: None
-        The name of the X variable
-        If a DataFrame is passed to fit and feature is None, feature
-        is selected as the column of the DataFrame.  There must be only
-        one column in the DataFrame.
+    columns : int, str, [int, int], [str, str], default: None
+        Determines what data is plotted in the joint plot and acts as a selection index
+        into the data passed to ``fit(X, y)``. This data therefore must be indexable by
+        the column type (e.g. an int for a numpy array or a string for a DataFrame).
 
-    target: string, default: None
-        The name of the Y variable
-        If target is None and a y value is passed to fit then the target
-        is selected from the target vector.
+        If None is specified then either both X and y must be 1D vectors and they will
+        be plotted against each other or X must be a 2D array with only 2 columns. If a
+        single index is specified then the data is indexed as ``X[columns]`` and plotted
+        jointly with the target variable, y. If two indices are specfied then they are
+        both selected from X, additionally in this case, if y is specified, then it is
+        used to plot the color of points.
 
[…]
