+++ tests/test_features/test_jointplot.py
@@ -1,139 +1,235 @@
 # tests.test_features.test_jointplot
-# Test the JointPlotVisualizer
+# Test the JointPlot Visualizer
 #
 # Author:   Prema Damodaran Roman
 # Created:  Mon Apr 10 21:00:54 2017 -0400
 #
-# Copyright (C) 2017 District Data Labs
+# Copyright (C) 2017 The scikit-yb developers.
 # For license information, see LICENSE.txt
 #
 # ID: test_jointplot.py [9e008b0] pdamodaran@users.noreply.github.com $
 
 """
-Test the JointPlotVisualizer.
+Test joint plot visualization methods.
 
 These tests work differently depending on what version of matplotlib is
-installed. If version 2.0.0 or greater is installed, then most tests will
-execute, otherwise most will skip and only the warning will be tested.
+installed. If version 2.0.2 or greater is installed, then most tests will
+execute, otherwise the histogram tests will skip and only the warning will
+be tested.
 """
 
 ##########################################################################
 ## Imports
 ##########################################################################
 
-import sys
 import pytest
-import warnings
 import numpy as np
-import matplotlib as mpl
-import matplotlib.pyplot as plt
 
-from tests.dataset import DatasetMixin
+from functools import partial
+from tests.dataset import Dataset
 from tests.base import VisualTestCase
 from yellowbrick.features.jointplot import *
+from yellowbrick.exceptions import YellowbrickValueError
+from sklearn.datasets import make_classification, make_regression
+
+try:
+    # Only available in Matplotlib >= 2.0.2
+    from mpl_toolkits.axes_grid1 import make_axes_locatable
+except ImportError:
+    make_axes_locatable = None
+
+try:
+    from unittest.mock import patch
+except ImportError:
+    from mock import patch
+
 
 ##########################################################################
-## JointPlotVisualizer Tests
+## Fixtures
 ##########################################################################
 
-# Determine version of matplotlib
-MPL_VERS_MAJ = int(mpl.__version__.split(".")[0])
+# Random numpy array generators
+rand1d = partial(np.random.rand, 120)
+rand2col = partial(np.random.rand, 120, 2)
+rand3col = partial(np.random.rand, 120, 3)
 
 
-class JointPlotTests(VisualTestCase, DatasetMixin):
-
-    X = np.array([1, 2, 3, 5, 8, 10])
-
-    y = np.array([1, 3, 6, 2, 9, 2])
-
-    def setUp(self):
-        self.concrete = self.load_data('concrete')
-
-    def tearDown(self):
-        self.concrete = None
-
-    @pytest.mark.skipif(MPL_VERS_MAJ > 1, reason="requires matplotlib 1.5.3 or less")
-    def test_warning(self):
-        """
-        Ensure that the jointplot warns if mpl version is < 2.0.0
-        """
-        # Note Python 3.2+ has a self.assertWarns ... but we need to be
-        # Python 2.7 compatible, so we're going to do this.
-        with warnings.catch_warnings(record=True) as w:
-            # Cause all warnings to always be triggered.
-            warnings.simplefilter("always")
-
-            # Trigger a warning.
-            JointPlotVisualizer()
-
-            # Ensure that a warning occurred
-            self.assertEqual(len(w), 1)
-            self.assertEqual(
-                str(w[-1].message),
-                "JointPlotVisualizer requires matplotlib major version 2 "
-                "or greater. Please upgrade."
-            )
-
-    @pytest.mark.xfail(
-        sys.platform == 'win32', reason="images not close on windows"
+@pytest.fixture(scope='class')
+def discrete(request):
+    """
+    Creates a simple 2-column dataset with a discrete target.
+    """
+    X, y = make_classification(
[…]
