### 2019-02-03 31bc784
added additional correlation measures
