### 2019-02-04 543168c
added none columns tests
