### 2019-02-05 f3c7c58
changes to get tests to pass
