### 2019-02-06 d9f4bed
Hotfix v0.9.1: matplotlib 3 support (#724)
This hotfix adds matplotlib3 support by requiring any version
of mpl except for 3.0.0 which had a backend bug that affected
Yellowbrick. Note that this hotfix includes changes to tests
that will need to be resolved when merging from develop.
