+++ yellowbrick/version.py
@@ -20,7 +20,7 @@ Maintains version and package information for deployment.
 __version_info__ = {
     'major': 0,
     'minor': 9,
-    'micro': 0,
+    'micro': 1,
     'releaselevel': 'final',
     'serial': 13,
 }
