### 2019-02-06 40f477a
completed image similarity tests
