### 2019-02-06 6ba070f
Alter ParallelCoordinates docs to include plot:: directives
1.  Changed code to make sure images were identical to those being replaced
