+++ docs/api/features/pcoords.rst
@@ -5,12 +5,15 @@ Parallel Coordinates
 
 Parallel coordinates is multi-dimensional feature visualization technique  where the vertical axis is duplicated horizontally for each feature. Instances are displayed as a single line segment drawn from each vertical axes to the location representing their value for that feature. This allows many dimensions to be visualized at once; in fact given infinite horizontal space (e.g. a scrolling window), technically an infinite number of dimensions can be displayed!
 
-Data scientists use this method to detect clusters of instances that have similar classes, and to note features that have high variance or different distributions. We can see this in action after first loading our occupancy classification dataset:
+Data scientists use this method to detect clusters of instances that have similar classes, and to note features that have high variance or different distributions. We can see this in action after first loading our occupancy classification dataset. Note that the visualization can be drawn with either the ``ParallelCoordinates`` visualizer or using the ``parallel_coordinates`` quick method:
 
-.. code:: python
+.. plot::
+
+    from yellowbrick.features import ParallelCoordinates
+    from yellowbrick.datasets import load_occupancy
 
     # Load the classification data set
-    data = load_data("occupancy")
+    X, y = load_occupancy()
 
     # Specify the features of interest and the classes of the target
     features = [
@@ -18,29 +21,18 @@ Data scientists use this method to detect clusters of instances that have simila
     ]
     classes = ["unoccupied", "occupied"]
 
-    # Extract the instances and target
-    X = data[features]
-    y = data.occupancy
-
-The visualization can be drawn with either the ``ParallelCoordinates`` visualizer or using the ``parallel_coordinates`` quick method:
-
-.. code:: python
-
-    from yellowbrick.features import ParallelCoordinates
-
     # Instantiate the visualizer
     visualizer = ParallelCoordinates(
-        classes=classes, features=features, sample=0.5, shuffle=True
+        classes=classes, features=features, sample=0.05, shuffle=True
     )
 
     # Fit and transform the data to the visualizer
     visualizer.fit_transform(X, y)
 
     # Finalize the title and axes then display the visualization
-    visualizer.poof()
+    visualizer.finalize()
+    visualizer.ax.grid(False)
 
-
-.. image:: images/parallel_coordinates.png
 
 By inspecting the visualization closely, we can see that the combination of transparency and overlap gives us the sense of groups of similar instances, sometimes referred to as "braids". If there are distinct braids of different classes, it suggests that there is enough separability that a classification algorithm might be able to discern between each class.
 
@@ -48,9 +40,19 @@ Unfortunately, as we inspect this class, we can see that the domain of each feat
 
 Normalization techniques can be directly applied to the visualizer without pre-transforming the data (though you could also do this) by using the ``normalize`` parameter. Several transformers are available; try using ``minmax``, ``minabs``, ``standard``, ``l1``, or ``l2`` normalization to change perspectives in the parallel coordinates as follows:
 
-.. code:: python
+.. plot::
 
     from yellowbrick.features import ParallelCoordinates
+    from yellowbrick.datasets import load_occupancy
+
+    # Load the classification data set
+    X, y = load_occupancy()
+
+    # Specify the features of interest and the classes of the target
+    features = [
+        "temperature", "relative humidity", "light", "C02", "humidity"
+    ]
+    classes = ["unoccupied", "occupied"]
 
     # Instantiate the visualizer
     visualizer = ParallelCoordinates(
@@ -60,9 +62,9 @@ Normalization techniques can be directly applied to the visualizer without pre-t
 
[…]
