### 2019-02-06 b9dd495
adjust tolerences for appveyor
