### 2019-02-07 d114568
Address @bbengfort PR Comments
1.  changed ``finalize`` and ``ax.grid()`` back to ``poof()``
2.  added note directive
3.  removed code for benchmark plot and put back in the benchmark static image.
Essentially, we put this code back into a `pcoord_benchmark.py` file and kept `.. image:: images/fast_parallel_coordinates_speedup.png` here
