+++ docs/api/features/pcoords.rst
@@ -5,7 +5,9 @@ Parallel Coordinates
 
 Parallel coordinates is multi-dimensional feature visualization technique  where the vertical axis is duplicated horizontally for each feature. Instances are displayed as a single line segment drawn from each vertical axes to the location representing their value for that feature. This allows many dimensions to be visualized at once; in fact given infinite horizontal space (e.g. a scrolling window), technically an infinite number of dimensions can be displayed!
 
-Data scientists use this method to detect clusters of instances that have similar classes, and to note features that have high variance or different distributions. We can see this in action after first loading our occupancy classification dataset. Note that the visualization can be drawn with either the ``ParallelCoordinates`` visualizer or using the ``parallel_coordinates`` quick method:
+Data scientists use this method to detect clusters of instances that have similar classes, and to note features that have high variance or different distributions. We can see this in action after first loading our occupancy classification dataset.
+
+.. note:: These visualizations can be produced with either the ``ParallelCoordinates`` visualizer or by using the ``parallel_coordinates`` quick method.
 
 .. plot::
 
@@ -30,8 +32,7 @@ Data scientists use this method to detect clusters of instances that have simila
     visualizer.fit_transform(X, y)
 
     # Finalize the title and axes then display the visualization
-    visualizer.finalize()
-    visualizer.ax.grid(False)
+    visualizer.poof()
 
 
 By inspecting the visualization closely, we can see that the combination of transparency and overlap gives us the sense of groups of similar instances, sometimes referred to as "braids". If there are distinct braids of different classes, it suggests that there is enough separability that a classification algorithm might be able to discern between each class.
@@ -62,8 +63,7 @@ Normalization techniques can be directly applied to the visualizer without pre-t
 
     # Fit the visualizer and display it
     visualizer.fit_transform(X, y)
-    visualizer.finalize()
-    visualizer.ax.grid(False)
+    visualizer.poof()
 
 
 Now we can see that each feature is in the range ``[-3, 3]`` where the mean of the feature is set to zero and each feature has a unit variance applied between ``[-1, 1]`` (because we're using the ``StandardScaler`` via the ``standard`` normalize parameter). This version of parallel coordinates gives us a much better sense of the distribution of the features and if any features are highly variable with respect to any one class.
@@ -96,80 +96,12 @@ The "fast" drawing mode vastly improves the performance of the parallel coordina
         oz = ParallelCoordinates(ax=axes[idx], fast=fast, title=title)
         oz.fit_transform(data.data, data.target)
         oz.finalize()
-        oz.ax.grid(False)
 
     plt.tight_layout()
 
 As you can see the "fast" drawing algorithm does not have the same build up of color density where instances of the same class intersect. Because there is only one line per class, there is only a darkening effect between classes. This can lead to a different interpretation of the plot, though it still may be effective for analytical purposes, particularly when you're plotting a lot of data. Needless to say, the performance benefits are dramatic:
 
-.. plot::
-    :include-source: False
-
-    import time
-    import matplotlib.pyplot as plt
-    from sklearn.datasets import load_iris
-    from yellowbrick.features import ParallelCoordinates
-    import pandas as pd
-
-    trials = 5
-    factors = np.arange(1, 11)
-
-    def pcoords_time(X, y, fast=True):
-        _, ax = plt.subplots()
-        oz = ParallelCoordinates(fast=fast, ax=ax)
-
-        start = time.time()
-        oz.fit_transform(X, y)
-        delta = time.time() - start
-
[…]
