### 2019-02-07 6938fb9
docstring updates
