### 2019-02-07 d2e2e2c
adds conventions for merging in PRs and a note about installing in editable mode for contributors
