### 2019-02-07 93736f1
reincorporates local branch merging convention
