### 2019-02-07 ed29ec5
updates to reflect practices used by other scikit yb devs
