### 2019-02-08 eb34e87
add close-figs context
