### 2019-02-09 289ee51
made suggested changes and updated with test functions as mentioned for prcurve
