### 2019-02-09 6288e3c
minor cleanups
