### 2019-02-09 2b1cfeb
fix CI tests
