### 2019-02-09 e1b7aaf
added instructions for remote forks and moved branching conventions to getting started section based on questions that have come up from contributors
