### 2019-02-10 3205b64
text and decision boundaries
