### 2019-02-10 17329d2
Add credit dataset to doc.  It is incomplete though
