+++ tests/test_text/test_dispersion.py
@@ -21,23 +21,28 @@ Tests for the dispersion plot text visualization
 import pytest
 
 from yellowbrick.exceptions import YellowbrickValueError
+from yellowbrick.datasets import load_hobbies
 from yellowbrick.text.dispersion import *
-from tests.dataset import DatasetMixin
 from tests.base import VisualTestCase
 import matplotlib.pyplot as plt
+
+##########################################################################
+## Data
+##########################################################################
+
+corpus = load_hobbies()
 
 ##########################################################################
 ## DispersionPlot Tests
 ##########################################################################
 
-class DispersionPlotTests(VisualTestCase, DatasetMixin):
+class DispersionPlotTests(VisualTestCase):
 
     def test_quick_method(self):
         """
         Assert no errors occur when using the qucik method
         """
         _, ax = plt.subplots()
-        corpus = self.load_data('hobbies')
 
         text = [doc.split() for doc in corpus.data]
         target_words = ['Game', 'player', 'score', 'oil', 'Man']
@@ -47,12 +52,10 @@ class DispersionPlotTests(VisualTestCase, DatasetMixin):
 
         self.assert_images_similar(ax=ax, tol=25)
 
-    def test_integrated_dispersionplot(self):
+    def test_integrated_dispersion_plot(self):
         """
         Assert no errors occur during DispersionPlot integration
         """
-        corpus = self.load_data('hobbies')
-
         text = [doc.split() for doc in corpus.data]
         target_words = ['Game', 'player', 'score', 'oil', 'Man']
 
@@ -62,13 +65,11 @@ class DispersionPlotTests(VisualTestCase, DatasetMixin):
 
         self.assert_images_similar(visualizer, tol=25)
 
-    def test_dispersionplot_ignore_case(self):
+    def test_dispersion_plot_ignore_case(self):
         """
         Assert no errors occur during DispersionPlot integration
         with ignore_case parameter turned on
         """
-        corpus = self.load_data('hobbies')
-
         text = [doc.split() for doc in corpus.data]
         target_words = ['Game', 'player', 'score', 'oil', 'Man']
 
@@ -78,14 +79,12 @@ class DispersionPlotTests(VisualTestCase, DatasetMixin):
 
         self.assert_images_similar(visualizer, tol=25)
 
-    def test_dispersionplot_generator_input(self):
+    def test_dispersion_plot_generator_input(self):
         """
         Assert no errors occur during dispersionPlot integration
         when the corpus' text type is a generator
         """
-        corpus = self.load_data('hobbies')
-
-        text = (doc.split() for doc in corpus.data)
+        text = [doc.split() for doc in corpus.data]
         target_words = ['Game', 'player', 'score', 'oil', 'Man']
 
         visualizer = DispersionPlot(target_words, ignore_case=True)
@@ -94,13 +93,11 @@ class DispersionPlotTests(VisualTestCase, DatasetMixin):
 
         self.assert_images_similar(visualizer, tol=25)
 
-    def test_dispersionplot_annotate_docs(self):
+    def test_dispersion_plot_annotate_docs(self):
         """
         Assert no errors occur during DispersionPlot integration
         with annotate_docs parameter turned on
         """
-        corpus = self.load_data('hobbies')
-
         text = [doc.split() for doc in corpus.data]
         target_words = ['girl', 'she', 'boy', 'he', 'man']
 
@@ -110,34 +107,28 @@ class DispersionPlotTests(VisualTestCase, DatasetMixin):
 
         self.assert_images_similar(visualizer, tol=25)
 
-    def test_dispersionplot_color_words_by_class(self):
+    def test_dispersion_plot_color_by_class(self):
         """
         Assert no errors occur during DispersionPlot integration
[…]
