### 2019-02-10 a715e29
updates frequency distribution tests to use new datasets module
