### 2019-02-10 7e79fd1
updates frequency tsne and umap tests to use new datasets module
