### 2019-02-10 9181861
removes DatasetMixin from umap tests
