### 2019-02-10 acd4750
Fix 'nofigs' option so that image is hidden
