### 2019-02-10 e70859a
first crack at a documentation readme for #450
