### 2019-02-10 98bd553
tiny fixes to make sure jointplot shows up properly in quickstart and autodoc
