### 2019-02-11 e172226
Add :Context: close-figs to elbow and importances plot directives
