+++ docs/quickstart.rst
@@ -152,7 +152,6 @@ The machine learning workflow is the art of creating *model selection triples*, 
    visualizer.poof()
 
 .. plot::
-    :context: close-figs
     :include-source: False
     :context: close-figs
     :alt: Rank2D of Bikeshare Features
@@ -180,7 +179,6 @@ In this chart, we see that the features ``temp`` and ``feelslike`` have a strong
     visualizer.poof()
 
 .. plot::
-    :context: close-figs
     :include-source: False
     :context: close-figs
     :alt: JointPlot of temp vs feelslike
@@ -220,7 +218,6 @@ At this point, we can train our model; let's fit a linear regression to our mode
     visualizer.poof()
 
 .. plot::
-    :context: close-figs
     :include-source: False
     :context: close-figs
     :alt: ResidualsPlot of a simple LinearRegression
@@ -263,7 +260,6 @@ Along with generating the residuals plot, we also measured the performance by "s
     visualizer.poof()
 
 .. plot::
-    :context: close-figs
     :include-source: False
     :context: close-figs
     :alt: AlphaSelection for L2 Regularization using RidgeCV
@@ -299,7 +295,6 @@ We can now train our final model and visualize it with the ``PredictionError`` v
     visualizer.poof()
 
 .. plot::
-    :context: close-figs
     :include-source: False
     :context: close-figs
     :alt: PredictionError for L2 Regularization using Ridge
