### 2019-02-11 84c80c7
update plot directive so that all the imports are added
