### 2019-02-11 c1a8a4c
Update code blocks in SilhouetteVisualizer docs
Updates the code block in the SilhouetteVisualizer documentation so that it can utilize the new plot directive from PR #446 to autogenerate the images.

Issue #687
