### 2019-02-11 329282b
Implement plot directive for ROCAUC
