+++ docs/api/classifier/rocauc.rst
@@ -9,38 +9,29 @@ The Receiver Operating Characteristic (ROC) is a measure of a classifier's predi
 
 This leads to another metric, area under the curve (AUC), which is a computation of the relationship between false positives and true positives. The higher the AUC, the better the model generally is. However, it is also important to inspect the "steepness" of the curve, as this describes the maximization of the true positive rate while minimizing the false positive rate.
 
-.. code:: python
+.. plot::
+    :context: close-figs
 
     from sklearn.model_selection import train_test_split
+    from yellowbrick.classifier import ROCAUC
+    from sklearn.linear_model import LogisticRegression
+    from yellowbrick.datasets import load_occupancy
 
     # Load the classification data set
-    data = load_data("occupancy")
+    X, y = load_occupancy()
 
-    # Specify the features of interest and the classes of the target
-    features = ["temperature", "relative humidity", "light", "C02", "humidity"]
+    # Specify the classes of the target
     classes = ["unoccupied", "occupied"]
-
-    # Extract the instances and target
-    X = data[features]
-    y = data.occupancy
 
     # Create the train and test data
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
-
-.. code:: python
-
-    from yellowbrick.classifier import ROCAUC
-    from sklearn.linear_model import LogisticRegression
 
     # Instantiate the visualizer with the classification model
     visualizer = ROCAUC(LogisticRegression(), classes=classes)
 
     visualizer.fit(X_train, y_train)  # Fit the training data to the visualizer
     visualizer.score(X_test, y_test)  # Evaluate the model on the test data
-    g = visualizer.poof()             # Draw/show/poof the data
-
-
-.. image:: images/rocauc_binary.png
+    visualizer.poof()             # Draw/show/poof the data
 
 
 .. warning::
@@ -58,37 +49,38 @@ Multi-class ROCAUC Curves
 Yellowbrick's ``ROCAUC`` Visualizer does allow for plotting multiclass classification curves.
 ROC curves are typically used in binary classification, and in fact the Scikit-Learn ``roc_curve`` metric is only able to perform metrics for binary classifiers. Yellowbrick addresses this by binarizing the output (per-class) or to use one-vs-rest (micro score) or one-vs-all (macro score) strategies of classification.
 
-.. code::
+.. plot::
+    :context: close-figs
+
+    from sklearn.model_selection import train_test_split
+    from yellowbrick.classifier import ROCAUC
+    from yellowbrick.datasets import load_game
+    from sklearn.linear_model import RidgeClassifier
+    from sklearn.preprocessing import OrdinalEncoder, LabelEncoder
 
     # Load multi-class classification dataset
-    game = load_game()
+    X, y = load_game()
 
     classes = ["win", "loss", "draw"]
 
     # Encode the non-numeric columns
-    game.replace({'loss':-1, 'draw':0, 'win':1, 'x':2, 'o':3, 'b':4}, inplace=True)
+    X = OrdinalEncoder().fit_transform(X)
 
-    # Extract the instances and target
-    X = game.iloc[:, game.columns != 'outcome']
-    y = game['outcome']
+    y = LabelEncoder().fit_transform(y)
 
     # Create the train and test data
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
-
-.. code::
-
-    from sklearn.linear_model import RidgeClassifier
 
     visualizer = ROCAUC(RidgeClassifier(), classes=classes)
 
     visualizer.fit(X_train, y_train)  # Fit the training data to the visualizer
     visualizer.score(X_test, y_test)  # Evaluate the model on the test data
-    g = visualizer.poof()             # Draw/show/poof the data
[…]
