### 2019-02-12 339b232
Update Intercluster Distance Maps documentation
Add new plot directive settings from PR #446 to Intercluster Distance Maps
Visualizer documentation. In addition, I deleted the original script
that had previously been used to generate the image in the documentation.

See also: #687
