### 2019-02-12 b646cf9
Fix parameter in Intercluster Distance docs
Originally, in my PR (#742) I had set the parameter in k in the Intercluster Distance
visualizer, similar to how k is set in KElbowVisualizer. And while it
still was able generate the image, I believe this is what caused the
test to fail.
