### 2019-02-12 23072d4
Rewrite warning for Index Error regarding binary classification
