### 2019-02-12 ac07bc8
Modify per Ben's comments
