### 2019-02-12 01e17b9
Address @rebeccabilbro concerns in PR
1.  Add intrepretation to figures (might have done a horrible job)
2.  Included second figure by removing :nofigs: and "include-source"
3.  Added introduction for Class Prediction Error
