### 2019-02-12 95c3b74
bump miniconda version
