### 2019-02-13 6b8eecd
attempt #2 with conda update
