### 2019-03-23 #1075 @arfon
One quick question @bbengfort - I notice there are a lot of contributors to this package, and your paper mentions a number of core contributors yet I only see two of you listed as authors on the JOSS paper. Is this correct?
