### 2019-03-23 #1075 @bbengfort
@mnarayan thank you so much for the review and the compliment!

@arfon we're really excited about this publication! The DOI of the official Yellowbrick archive (with all versions) is [10.5281/zenodo.1488364](https://zenodo.org/record/1488364). As for your question about the authors - @rebeccabilbro and I are the primary authors of Yellowbrick and the maintainers. After discussing with our other core contributors we feel that the easiest and most equitable thing for everyone is if it's just the two of us listed as authors on the paper and our core contributors are mentioned in the acknowledgements. We are proud of the number of contributors we've had to the package, and on every release we list everyone who contributed as recognition.

Thanks again for all your hard work on JOSS!
