### 2019-03-23 #1075 @rebeccabilbro
Thank you so much @mnarayan for your review &mdash; we are excited to get the word out about visual diagnostics for model selection via JOSS!! Sincere thanks also to @arfon for shepherding our paper through the process; we are deeply grateful to you both for volunteering your time and energies to this important work!
