### 2019-03-24 #1075 @bbengfort
@arfon could we update our archive DOI to 10.5281/zenodo.1206239 please? This will resolve to the latest version of Yellowbrick, whereas the DOI I supplied before will only resolve to v0.9.

The DOI for the Interactive Optimization paper is correct, I'm updating the bibtex now!
