### 2019-03-24 #1075 @bbengfort
@arfon the bibtex has been updated in the Yellowbrick repo.
