### 2019-03-24 #1075 @arfon
@mnarayan - many thanks for your review here ✨

@bbengfort @rebeccabilbro - your paper is now accepted into JOSS :zap::rocket::boom:
